% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{AccessPolicy}
\alias{AccessPolicy}
\title{YouTube Data API Objects 
Supports core YouTube features, such as uploading videos, creating and managing playlists, searching for content, and much more.}
\usage{
AccessPolicy(allowed = NULL, exception = NULL)
}
\arguments{
\item{allowed}{The value of allowed indicates whether the access to the policy is allowed or denied by default}

\item{exception}{A list of region codes that identify countries where the default policy do not apply}
}
\value{
AccessPolicy object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:50:58
filename: /Users/mark/dev/R/autoGoogleAPI/googleyoutubev3.auto/R/youtube_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AccessPolicy Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Rights management policy for YouTube resources.
}
