% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{activities.insert}
\alias{activities.insert}
\title{Posts a bulletin for a specific channel. (The user submitting the request must be authorized to act on the channel's behalf.)Note: Even though an activity resource can contain information about actions like a user rating a video or marking a video as a favorite, you need to use other API methods to generate those activity resources. For example, you would use the API's videos.rate() method to rate a video and the playlistItems.insert() method to mark a video as a favorite.}
\usage{
activities.insert(Activity, part)
}
\arguments{
\item{Activity}{The \link{Activity} object to pass to this method}

\item{part}{The part parameter serves two purposes in this operation}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other Activity functions: \code{\link{Activity}}
}
