% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{activities.list}
\alias{activities.list}
\title{Returns a list of channel activity events that match the request criteria. For example, you can retrieve events associated with a particular channel, events associated with the user's subscriptions and Google+ friends, or the YouTube home page feed, which is customized for each user.}
\usage{
activities.list(part, channelId = NULL, home = NULL, maxResults = NULL,
  mine = NULL, pageToken = NULL, publishedAfter = NULL,
  publishedBefore = NULL, regionCode = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more activity resource properties that the API response will include}

\item{channelId}{The channelId parameter specifies a unique YouTube channel ID}

\item{home}{Set this parameter's value to true to retrieve the activity feed that displays on the YouTube home page for the currently authenticated user}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{mine}{Set this parameter's value to true to retrieve a feed of the authenticated user's activities}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}

\item{publishedAfter}{The publishedAfter parameter specifies the earliest date and time that an activity could have occurred for that activity to be included in the API response}

\item{publishedBefore}{The publishedBefore parameter specifies the date and time before which an activity must have occurred for that activity to be included in the API response}

\item{regionCode}{The regionCode parameter instructs the API to return results for the specified country}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

