% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetails}
\alias{ActivityContentDetails}
\title{ActivityContentDetails Object}
\usage{
ActivityContentDetails(bulletin = NULL, channelItem = NULL,
  comment = NULL, favorite = NULL, like = NULL, playlistItem = NULL,
  promotedItem = NULL, recommendation = NULL, social = NULL,
  subscription = NULL, upload = NULL)
}
\arguments{
\item{bulletin}{The bulletin object contains details about a channel bulletin post}

\item{channelItem}{The channelItem object contains details about a resource which was added to a channel}

\item{comment}{The comment object contains information about a resource that received a comment}

\item{favorite}{The favorite object contains information about a video that was marked as a favorite video}

\item{like}{The like object contains information about a resource that received a positive (like) rating}

\item{playlistItem}{The playlistItem object contains information about a new playlist item}

\item{promotedItem}{The promotedItem object contains details about a resource which is being promoted}

\item{recommendation}{The recommendation object contains information about a recommended resource}

\item{social}{The social object contains details about a social network post}

\item{subscription}{The subscription object contains information about a channel that a user subscribed to}

\item{upload}{The upload object contains information about the uploaded video}
}
\value{
ActivityContentDetails object
}
\description{
ActivityContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about the content of an activity: the video that was shared, the channel that was subscribed to, etc.
}
