% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsBulletin}
\alias{ActivityContentDetailsBulletin}
\title{ActivityContentDetailsBulletin Object}
\usage{
ActivityContentDetailsBulletin(resourceId = NULL)
}
\arguments{
\item{resourceId}{The resourceId object contains information that identifies the resource associated with a bulletin post}
}
\value{
ActivityContentDetailsBulletin object
}
\description{
ActivityContentDetailsBulletin Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about a channel bulletin post.
}

