% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsChannelItem}
\alias{ActivityContentDetailsChannelItem}
\title{ActivityContentDetailsChannelItem Object}
\usage{
ActivityContentDetailsChannelItem(resourceId = NULL)
}
\arguments{
\item{resourceId}{The resourceId object contains information that identifies the resource that was added to the channel}
}
\value{
ActivityContentDetailsChannelItem object
}
\description{
ActivityContentDetailsChannelItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about a resource which was added to a channel.
}

