% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsComment}
\alias{ActivityContentDetailsComment}
\title{ActivityContentDetailsComment Object}
\usage{
ActivityContentDetailsComment(resourceId = NULL)
}
\arguments{
\item{resourceId}{The resourceId object contains information that identifies the resource associated with the comment}
}
\value{
ActivityContentDetailsComment object
}
\description{
ActivityContentDetailsComment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a resource that received a comment.
}
