% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsFavorite}
\alias{ActivityContentDetailsFavorite}
\title{ActivityContentDetailsFavorite Object}
\usage{
ActivityContentDetailsFavorite(resourceId = NULL)
}
\arguments{
\item{resourceId}{The resourceId object contains information that identifies the resource that was marked as a favorite}
}
\value{
ActivityContentDetailsFavorite object
}
\description{
ActivityContentDetailsFavorite Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a video that was marked as a favorite video.
}
