% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsLike}
\alias{ActivityContentDetailsLike}
\title{ActivityContentDetailsLike Object}
\usage{
ActivityContentDetailsLike(resourceId = NULL)
}
\arguments{
\item{resourceId}{The resourceId object contains information that identifies the rated resource}
}
\value{
ActivityContentDetailsLike object
}
\description{
ActivityContentDetailsLike Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a resource that received a positive (like) rating.
}

