% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsPlaylistItem}
\alias{ActivityContentDetailsPlaylistItem}
\title{ActivityContentDetailsPlaylistItem Object}
\usage{
ActivityContentDetailsPlaylistItem(playlistId = NULL, playlistItemId = NULL,
  resourceId = NULL)
}
\arguments{
\item{playlistId}{The value that YouTube uses to uniquely identify the playlist}

\item{playlistItemId}{ID of the item within the playlist}

\item{resourceId}{The resourceId object contains information about the resource that was added to the playlist}
}
\value{
ActivityContentDetailsPlaylistItem object
}
\description{
ActivityContentDetailsPlaylistItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a new playlist item.
}
