% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsPromotedItem}
\alias{ActivityContentDetailsPromotedItem}
\title{ActivityContentDetailsPromotedItem Object}
\usage{
ActivityContentDetailsPromotedItem(adTag = NULL, clickTrackingUrl = NULL,
  creativeViewUrl = NULL, ctaType = NULL, customCtaButtonText = NULL,
  descriptionText = NULL, destinationUrl = NULL, forecastingUrl = NULL,
  impressionUrl = NULL, videoId = NULL)
}
\arguments{
\item{adTag}{The URL the client should fetch to request a promoted item}

\item{clickTrackingUrl}{The URL the client should ping to indicate that the user clicked through on this promoted item}

\item{creativeViewUrl}{The URL the client should ping to indicate that the user was shown this promoted item}

\item{ctaType}{The type of call-to-action, a message to the user indicating action that can be taken}

\item{customCtaButtonText}{The custom call-to-action button text}

\item{descriptionText}{The text description to accompany the promoted item}

\item{destinationUrl}{The URL the client should direct the user to, if the user chooses to visit the advertiser's website}

\item{forecastingUrl}{The list of forecasting URLs}

\item{impressionUrl}{The list of impression URLs}

\item{videoId}{The ID that YouTube uses to uniquely identify the promoted video}
}
\value{
ActivityContentDetailsPromotedItem object
}
\description{
ActivityContentDetailsPromotedItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about a resource which is being promoted.
}
