% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsRecommendation}
\alias{ActivityContentDetailsRecommendation}
\title{ActivityContentDetailsRecommendation Object}
\usage{
ActivityContentDetailsRecommendation(reason = NULL, resourceId = NULL,
  seedResourceId = NULL)
}
\arguments{
\item{reason}{The reason that the resource is recommended to the user}

\item{resourceId}{The resourceId object contains information that identifies the recommended resource}

\item{seedResourceId}{The seedResourceId object contains information about the resource that caused the recommendation}
}
\value{
ActivityContentDetailsRecommendation object
}
\description{
ActivityContentDetailsRecommendation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information that identifies the recommended resource.
}
