% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsSocial}
\alias{ActivityContentDetailsSocial}
\title{ActivityContentDetailsSocial Object}
\usage{
ActivityContentDetailsSocial(author = NULL, imageUrl = NULL,
  referenceUrl = NULL, resourceId = NULL, type = NULL)
}
\arguments{
\item{author}{The author of the social network post}

\item{imageUrl}{An image of the post's author}

\item{referenceUrl}{The URL of the social network post}

\item{resourceId}{The resourceId object encapsulates information that identifies the resource associated with a social network post}

\item{type}{The name of the social network}
}
\value{
ActivityContentDetailsSocial object
}
\description{
ActivityContentDetailsSocial Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about a social network post.
}

