% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsSubscription}
\alias{ActivityContentDetailsSubscription}
\title{ActivityContentDetailsSubscription Object}
\usage{
ActivityContentDetailsSubscription(resourceId = NULL)
}
\arguments{
\item{resourceId}{The resourceId object contains information that identifies the resource that the user subscribed to}
}
\value{
ActivityContentDetailsSubscription object
}
\description{
ActivityContentDetailsSubscription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a channel that a user subscribed to.
}

