% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivityContentDetailsUpload}
\alias{ActivityContentDetailsUpload}
\title{ActivityContentDetailsUpload Object}
\usage{
ActivityContentDetailsUpload(videoId = NULL)
}
\arguments{
\item{videoId}{The ID that YouTube uses to uniquely identify the uploaded video}
}
\value{
ActivityContentDetailsUpload object
}
\description{
ActivityContentDetailsUpload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the uploaded video.
}

