% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ActivitySnippet}
\alias{ActivitySnippet}
\title{ActivitySnippet Object}
\usage{
ActivitySnippet(channelId = NULL, channelTitle = NULL, description = NULL,
  groupId = NULL, publishedAt = NULL, thumbnails = NULL, title = NULL,
  type = NULL)
}
\arguments{
\item{channelId}{The ID that YouTube uses to uniquely identify the channel associated with the activity}

\item{channelTitle}{Channel title for the channel responsible for this activity}

\item{description}{The description of the resource primarily associated with the activity}

\item{groupId}{The group ID associated with the activity}

\item{publishedAt}{The date and time that the video was uploaded}

\item{thumbnails}{A map of thumbnail images associated with the resource that is primarily associated with the activity}

\item{title}{The title of the resource primarily associated with the activity}

\item{type}{The type of activity that the resource describes}
}
\value{
ActivitySnippet object
}
\description{
ActivitySnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about an activity, including title, description, thumbnails, activity type and group.
}

