% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Caption}
\alias{Caption}
\title{Caption Object}
\usage{
Caption(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the caption track}

\item{snippet}{The snippet object contains basic details about the caption}
}
\value{
Caption object
}
\description{
Caption Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A caption resource represents a YouTube caption track. A caption track is associated with exactly one YouTube video.
}
\seealso{
Other Caption functions: \code{\link{captions.insert}},
  \code{\link{captions.update}}
}
