% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{CaptionListResponse}
\alias{CaptionListResponse}
\title{CaptionListResponse Object}
\usage{
CaptionListResponse(etag = NULL, eventId = NULL, items = NULL,
  visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of captions that match the request criteria}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
CaptionListResponse object
}
\description{
CaptionListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
