% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{captions.delete}
\alias{captions.delete}
\title{Deletes a specified caption track.}
\usage{
captions.delete(id, onBehalfOf = NULL, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{id}{The id parameter identifies the caption track that is being deleted}

\item{onBehalfOf}{ID of the Google+ Page for the channel that the request is be on behalf of}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
