% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{captions.download}
\alias{captions.download}
\title{Downloads a caption track. The caption track is returned in its original format unless the request specifies a value for the tfmt parameter and in its original language unless the request specifies a value for the tlang parameter.}
\usage{
captions.download(id, onBehalfOf = NULL, onBehalfOfContentOwner = NULL,
  tfmt = NULL, tlang = NULL)
}
\arguments{
\item{id}{The id parameter identifies the caption track that is being retrieved}

\item{onBehalfOf}{ID of the Google+ Page for the channel that the request is be on behalf of}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{tfmt}{The tfmt parameter specifies that the caption track should be returned in a specific format}

\item{tlang}{The tlang parameter specifies that the API response should return a translation of the specified caption track}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

