% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{captions.insert}
\alias{captions.insert}
\title{Uploads a caption track.}
\usage{
captions.insert(Caption, part, onBehalfOf = NULL,
  onBehalfOfContentOwner = NULL, sync = NULL)
}
\arguments{
\item{Caption}{The \link{Caption} object to pass to this method}

\item{part}{The part parameter specifies the caption resource parts that the API response will include}

\item{onBehalfOf}{ID of the Google+ Page for the channel that the request is be on behalf of}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{sync}{The sync parameter indicates whether YouTube should automatically synchronize the caption file with the audio track of the video}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other Caption functions: \code{\link{Caption}},
  \code{\link{captions.update}}
}

