% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{captions.list}
\alias{captions.list}
\title{Returns a list of caption tracks that are associated with a specified video. Note that the API response does not contain the actual captions and that the captions.download method provides the ability to retrieve a caption track.}
\usage{
captions.list(part, videoId, id = NULL, onBehalfOf = NULL,
  onBehalfOfContentOwner = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more caption resource parts that the API response will include}

\item{videoId}{The videoId parameter specifies the YouTube video ID of the video for which the API should return caption tracks}

\item{id}{The id parameter specifies a comma-separated list of IDs that identify the caption resources that should be retrieved}

\item{onBehalfOf}{ID of the Google+ Page for the channel that the request is on behalf of}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

