% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{CaptionSnippet}
\alias{CaptionSnippet}
\title{CaptionSnippet Object}
\usage{
CaptionSnippet(audioTrackType = NULL, failureReason = NULL,
  isAutoSynced = NULL, isCC = NULL, isDraft = NULL, isEasyReader = NULL,
  isLarge = NULL, language = NULL, lastUpdated = NULL, name = NULL,
  status = NULL, trackKind = NULL, videoId = NULL)
}
\arguments{
\item{audioTrackType}{The type of audio track associated with the caption track}

\item{failureReason}{The reason that YouTube failed to process the caption track}

\item{isAutoSynced}{Indicates whether YouTube synchronized the caption track to the audio track in the video}

\item{isCC}{Indicates whether the track contains closed captions for the deaf and hard of hearing}

\item{isDraft}{Indicates whether the caption track is a draft}

\item{isEasyReader}{Indicates whether caption track is formatted for 'easy reader,' meaning it is at a third-grade level for language learners}

\item{isLarge}{Indicates whether the caption track uses large text for the vision-impaired}

\item{language}{The language of the caption track}

\item{lastUpdated}{The date and time when the caption track was last updated}

\item{name}{The name of the caption track}

\item{status}{The caption track's status}

\item{trackKind}{The caption track's type}

\item{videoId}{The ID that YouTube uses to uniquely identify the video associated with the caption track}
}
\value{
CaptionSnippet object
}
\description{
CaptionSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a caption track, such as its language and name.
}

