% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{CdnSettings}
\alias{CdnSettings}
\title{CdnSettings Object}
\usage{
CdnSettings(format = NULL, frameRate = NULL, ingestionInfo = NULL,
  ingestionType = NULL, resolution = NULL)
}
\arguments{
\item{format}{The format of the video stream that you are sending to Youtube}

\item{frameRate}{The frame rate of the inbound video data}

\item{ingestionInfo}{The ingestionInfo object contains information that YouTube provides that you need to transmit your RTMP or HTTP stream to YouTube}

\item{ingestionType}{The method or protocol used to transmit the video stream}

\item{resolution}{The resolution of the inbound video data}
}
\value{
CdnSettings object
}
\description{
CdnSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Brief description of the live stream cdn settings.
}

