% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Channel.localizations}
\alias{Channel.localizations}
\title{Channel.localizations Object}
\usage{
Channel.localizations()
}
\value{
Channel.localizations object
}
\description{
Channel.localizations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Localizations for different languages
}
\seealso{
Other Channel functions: \code{\link{Channel}},
  \code{\link{channels.update}}
}
