% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Channel}
\alias{Channel}
\title{Channel Object}
\usage{
Channel(Channel.localizations = NULL, auditDetails = NULL,
  brandingSettings = NULL, contentDetails = NULL,
  contentOwnerDetails = NULL, conversionPings = NULL, etag = NULL,
  id = NULL, invideoPromotion = NULL, localizations = NULL,
  snippet = NULL, statistics = NULL, status = NULL, topicDetails = NULL)
}
\arguments{
\item{Channel.localizations}{The \link{Channel.localizations} object or list of objects}

\item{auditDetails}{The auditionDetails object encapsulates channel data that is relevant for YouTube Partners during the audition process}

\item{brandingSettings}{The brandingSettings object encapsulates information about the branding of the channel}

\item{contentDetails}{The contentDetails object encapsulates information about the channel's content}

\item{contentOwnerDetails}{The contentOwnerDetails object encapsulates channel data that is relevant for YouTube Partners linked with the channel}

\item{conversionPings}{The conversionPings object encapsulates information about conversion pings that need to be respected by the channel}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the channel}

\item{invideoPromotion}{The invideoPromotion object encapsulates information about promotion campaign associated with the channel}

\item{localizations}{Localizations for different languages}

\item{snippet}{The snippet object contains basic details about the channel, such as its title, description, and thumbnail images}

\item{statistics}{The statistics object encapsulates statistics for the channel}

\item{status}{The status object encapsulates information about the privacy status of the channel}

\item{topicDetails}{The topicDetails object encapsulates information about Freebase topics associated with the channel}
}
\value{
Channel object
}
\description{
Channel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A channel resource contains information about a YouTube channel.
}
\seealso{
Other Channel functions: \code{\link{Channel.localizations}},
  \code{\link{channels.update}}
}
