% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelAuditDetails}
\alias{ChannelAuditDetails}
\title{ChannelAuditDetails Object}
\usage{
ChannelAuditDetails(communityGuidelinesGoodStanding = NULL,
  contentIdClaimsGoodStanding = NULL, copyrightStrikesGoodStanding = NULL,
  overallGoodStanding = NULL)
}
\arguments{
\item{communityGuidelinesGoodStanding}{Whether or not the channel respects the community guidelines}

\item{contentIdClaimsGoodStanding}{Whether or not the channel has any unresolved claims}

\item{copyrightStrikesGoodStanding}{Whether or not the channel has any copyright strikes}

\item{overallGoodStanding}{Describes the general state of the channel}
}
\value{
ChannelAuditDetails object
}
\description{
ChannelAuditDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The auditDetails object encapsulates channel data that is relevant for YouTube Partners during the audit process.
}
