% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelBannerResource}
\alias{ChannelBannerResource}
\title{ChannelBannerResource Object}
\usage{
ChannelBannerResource(etag = NULL, url = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{url}{The URL of this banner image}
}
\value{
ChannelBannerResource object
}
\description{
ChannelBannerResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A channel banner returned as the response to a channel_banner.insert call.
}
\seealso{
Other ChannelBannerResource functions: \code{\link{channelBanners.insert}}
}

