% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{channelBanners.insert}
\alias{channelBanners.insert}
\title{Uploads a channel banner image to YouTube. This method represents the first two steps in a three-step process to update the banner image for a channel:- Call the channelBanners.insert method to upload the binary image data to YouTube. The image must have a 16:9 aspect ratio and be at least 2120x1192 pixels.- Extract the url property's value from the response that the API returns for step 1.- Call the channels.update method to update the channel's branding settings. Set the brandingSettings.image.bannerExternalUrl property's value to the URL obtained in step 2.}
\usage{
channelBanners.insert(ChannelBannerResource, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{ChannelBannerResource}{The \link{ChannelBannerResource} object to pass to this method}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.upload
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.upload)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other ChannelBannerResource functions: \code{\link{ChannelBannerResource}}
}

