% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelBrandingSettings}
\alias{ChannelBrandingSettings}
\title{ChannelBrandingSettings Object}
\usage{
ChannelBrandingSettings(channel = NULL, hints = NULL, image = NULL,
  watch = NULL)
}
\arguments{
\item{channel}{Branding properties for the channel view}

\item{hints}{Additional experimental branding properties}

\item{image}{Branding properties for branding images}

\item{watch}{Branding properties for the watch page}
}
\value{
ChannelBrandingSettings object
}
\description{
ChannelBrandingSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Branding properties of a YouTube channel.
}

