% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelContentDetails}
\alias{ChannelContentDetails}
\title{ChannelContentDetails Object}
\usage{
ChannelContentDetails(ChannelContentDetails.relatedPlaylists = NULL,
  googlePlusUserId = NULL, relatedPlaylists = NULL)
}
\arguments{
\item{ChannelContentDetails.relatedPlaylists}{The \link{ChannelContentDetails.relatedPlaylists} object or list of objects}

\item{googlePlusUserId}{The googlePlusUserId object identifies the Google+ profile ID associated with this channel}

\item{relatedPlaylists}{No description}
}
\value{
ChannelContentDetails object
}
\description{
ChannelContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about the content of a channel.
}
\seealso{
Other ChannelContentDetails functions: \code{\link{ChannelContentDetails.relatedPlaylists}}
}
