% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelContentDetails.relatedPlaylists}
\alias{ChannelContentDetails.relatedPlaylists}
\title{ChannelContentDetails.relatedPlaylists Object}
\usage{
ChannelContentDetails.relatedPlaylists(favorites = NULL, likes = NULL,
  uploads = NULL, watchHistory = NULL, watchLater = NULL)
}
\arguments{
\item{favorites}{The ID of the playlist that contains the channel's favorite videos}

\item{likes}{The ID of the playlist that contains the channel's liked videos}

\item{uploads}{The ID of the playlist that contains the channel's uploaded videos}

\item{watchHistory}{The ID of the playlist that contains the channel's watch history}

\item{watchLater}{The ID of the playlist that contains the channel's watch later playlist}
}
\value{
ChannelContentDetails.relatedPlaylists object
}
\description{
ChannelContentDetails.relatedPlaylists Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ChannelContentDetails functions: \code{\link{ChannelContentDetails}}
}
