% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelContentOwnerDetails}
\alias{ChannelContentOwnerDetails}
\title{ChannelContentOwnerDetails Object}
\usage{
ChannelContentOwnerDetails(contentOwner = NULL, timeLinked = NULL)
}
\arguments{
\item{contentOwner}{The ID of the content owner linked to the channel}

\item{timeLinked}{The date and time of when the channel was linked to the content owner}
}
\value{
ChannelContentOwnerDetails object
}
\description{
ChannelContentOwnerDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The contentOwnerDetails object encapsulates channel data that is relevant for YouTube Partners linked with the channel.
}
