% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelConversionPing}
\alias{ChannelConversionPing}
\title{ChannelConversionPing Object}
\usage{
ChannelConversionPing(context = NULL, conversionUrl = NULL)
}
\arguments{
\item{context}{Defines the context of the ping}

\item{conversionUrl}{The url (without the schema) that the player shall send the ping to}
}
\value{
ChannelConversionPing object
}
\description{
ChannelConversionPing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pings that the app shall fire (authenticated by biscotti cookie). Each ping has a context, in which the app must fire the ping, and a url identifying the ping.
}

