% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelConversionPings}
\alias{ChannelConversionPings}
\title{ChannelConversionPings Object}
\usage{
ChannelConversionPings(pings = NULL)
}
\arguments{
\item{pings}{Pings that the app shall fire (authenticated by biscotti cookie)}
}
\value{
ChannelConversionPings object
}
\description{
ChannelConversionPings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The conversionPings object encapsulates information about conversion pings that need to be respected by the channel.
}
