% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelLocalization}
\alias{ChannelLocalization}
\title{ChannelLocalization Object}
\usage{
ChannelLocalization(description = NULL, title = NULL)
}
\arguments{
\item{description}{The localized strings for channel's description}

\item{title}{The localized strings for channel's title}
}
\value{
ChannelLocalization object
}
\description{
ChannelLocalization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Channel localization setting
}

