% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelProfileDetails}
\alias{ChannelProfileDetails}
\title{ChannelProfileDetails Object}
\usage{
ChannelProfileDetails(channelId = NULL, channelUrl = NULL,
  displayName = NULL, profileImageUrl = NULL)
}
\arguments{
\item{channelId}{The YouTube channel ID}

\item{channelUrl}{The channel's URL}

\item{displayName}{The channel's display name}

\item{profileImageUrl}{The channels's avatar URL}
}
\value{
ChannelProfileDetails object
}
\description{
ChannelProfileDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
