% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{channels.list}
\alias{channels.list}
\title{Returns a collection of zero or more channel resources that match the request criteria.}
\usage{
channels.list(part, categoryId = NULL, forUsername = NULL, hl = NULL,
  id = NULL, managedByMe = NULL, maxResults = NULL, mine = NULL,
  mySubscribers = NULL, onBehalfOfContentOwner = NULL, pageToken = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more channel resource properties that the API response will include}

\item{categoryId}{The categoryId parameter specifies a YouTube guide category, thereby requesting YouTube channels associated with that category}

\item{forUsername}{The forUsername parameter specifies a YouTube username, thereby requesting the channel associated with that username}

\item{hl}{The hl parameter should be used for filter out the properties that are not in the given language}

\item{id}{The id parameter specifies a comma-separated list of the YouTube channel ID(s) for the resource(s) that are being retrieved}

\item{managedByMe}{Note: This parameter is intended exclusively for YouTube content partners}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{mine}{Set this parameter's value to true to instruct the API to only return channels owned by the authenticated user}

\item{mySubscribers}{Use the subscriptions}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
\item https://www.googleapis.com/auth/youtubepartner-channel-audit
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner, https://www.googleapis.com/auth/youtubepartner-channel-audit)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

