% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{channels.update}
\alias{channels.update}
\title{Updates a channel's metadata. Note that this method currently only supports updates to the channel resource's brandingSettings and invideoPromotion objects and their child properties.}
\usage{
channels.update(Channel, part, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{Channel}{The \link{Channel} object to pass to this method}

\item{part}{The part parameter serves two purposes in this operation}

\item{onBehalfOfContentOwner}{The onBehalfOfContentOwner parameter indicates that the authenticated user is acting on behalf of the content owner specified in the parameter value}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other Channel functions: \code{\link{Channel.localizations}},
  \code{\link{Channel}}
}

