% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSection.localizations}
\alias{ChannelSection.localizations}
\title{ChannelSection.localizations Object}
\usage{
ChannelSection.localizations()
}
\value{
ChannelSection.localizations object
}
\description{
ChannelSection.localizations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Localizations for different languages
}
\seealso{
Other ChannelSection functions: \code{\link{ChannelSection}},
  \code{\link{channelSections.insert}},
  \code{\link{channelSections.update}}
}

