% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSection}
\alias{ChannelSection}
\title{ChannelSection Object}
\usage{
ChannelSection(ChannelSection.localizations = NULL, contentDetails = NULL,
  etag = NULL, id = NULL, localizations = NULL, snippet = NULL,
  targeting = NULL)
}
\arguments{
\item{ChannelSection.localizations}{The \link{ChannelSection.localizations} object or list of objects}

\item{contentDetails}{The contentDetails object contains details about the channel section content, such as a list of playlists or channels featured in the section}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the channel section}

\item{localizations}{Localizations for different languages}

\item{snippet}{The snippet object contains basic details about the channel section, such as its type, style and title}

\item{targeting}{The targeting object contains basic targeting settings about the channel section}
}
\value{
ChannelSection object
}
\description{
ChannelSection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ChannelSection functions: \code{\link{ChannelSection.localizations}},
  \code{\link{channelSections.insert}},
  \code{\link{channelSections.update}}
}
