% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSectionContentDetails}
\alias{ChannelSectionContentDetails}
\title{ChannelSectionContentDetails Object}
\usage{
ChannelSectionContentDetails(channels = NULL, playlists = NULL)
}
\arguments{
\item{channels}{The channel ids for type multiple_channels}

\item{playlists}{The playlist ids for type single_playlist and multiple_playlists}
}
\value{
ChannelSectionContentDetails object
}
\description{
ChannelSectionContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about a channelsection, including playlists and channels.
}
