% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSectionListResponse}
\alias{ChannelSectionListResponse}
\title{ChannelSectionListResponse Object}
\usage{
ChannelSectionListResponse(etag = NULL, eventId = NULL, items = NULL,
  visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of ChannelSections that match the request criteria}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
ChannelSectionListResponse object
}
\description{
ChannelSectionListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
