% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSectionLocalization}
\alias{ChannelSectionLocalization}
\title{ChannelSectionLocalization Object}
\usage{
ChannelSectionLocalization(title = NULL)
}
\arguments{
\item{title}{The localized strings for channel section's title}
}
\value{
ChannelSectionLocalization object
}
\description{
ChannelSectionLocalization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ChannelSection localization setting
}

