% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSectionSnippet}
\alias{ChannelSectionSnippet}
\title{ChannelSectionSnippet Object}
\usage{
ChannelSectionSnippet(channelId = NULL, defaultLanguage = NULL,
  localized = NULL, position = NULL, style = NULL, title = NULL,
  type = NULL)
}
\arguments{
\item{channelId}{The ID that YouTube uses to uniquely identify the channel that published the channel section}

\item{defaultLanguage}{The language of the channel section's default title and description}

\item{localized}{Localized title, read-only}

\item{position}{The position of the channel section in the channel}

\item{style}{The style of the channel section}

\item{title}{The channel section's title for multiple_playlists and multiple_channels}

\item{type}{The type of the channel section}
}
\value{
ChannelSectionSnippet object
}
\description{
ChannelSectionSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a channel section, including title, style and position.
}

