% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSectionTargeting}
\alias{ChannelSectionTargeting}
\title{ChannelSectionTargeting Object}
\usage{
ChannelSectionTargeting(countries = NULL, languages = NULL,
  regions = NULL)
}
\arguments{
\item{countries}{The country the channel section is targeting}

\item{languages}{The language the channel section is targeting}

\item{regions}{The region the channel section is targeting}
}
\value{
ChannelSectionTargeting object
}
\description{
ChannelSectionTargeting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ChannelSection targeting setting.
}

