% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSettings}
\alias{ChannelSettings}
\title{ChannelSettings Object}
\usage{
ChannelSettings(country = NULL, defaultLanguage = NULL, defaultTab = NULL,
  description = NULL, featuredChannelsTitle = NULL,
  featuredChannelsUrls = NULL, keywords = NULL, moderateComments = NULL,
  profileColor = NULL, showBrowseView = NULL, showRelatedChannels = NULL,
  title = NULL, trackingAnalyticsAccountId = NULL,
  unsubscribedTrailer = NULL)
}
\arguments{
\item{country}{The country of the channel}

\item{defaultLanguage}{No description}

\item{defaultTab}{Which content tab users should see when viewing the channel}

\item{description}{Specifies the channel description}

\item{featuredChannelsTitle}{Title for the featured channels tab}

\item{featuredChannelsUrls}{The list of featured channels}

\item{keywords}{Lists keywords associated with the channel, comma-separated}

\item{moderateComments}{Whether user-submitted comments left on the channel page need to be approved by the channel owner to be publicly visible}

\item{profileColor}{A prominent color that can be rendered on this channel page}

\item{showBrowseView}{Whether the tab to browse the videos should be displayed}

\item{showRelatedChannels}{Whether related channels should be proposed}

\item{title}{Specifies the channel title}

\item{trackingAnalyticsAccountId}{The ID for a Google Analytics account to track and measure traffic to the channels}

\item{unsubscribedTrailer}{The trailer of the channel, for users that are not subscribers}
}
\value{
ChannelSettings object
}
\description{
ChannelSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Branding properties for the channel view.
}
