% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelSnippet}
\alias{ChannelSnippet}
\title{ChannelSnippet Object}
\usage{
ChannelSnippet(country = NULL, customUrl = NULL, defaultLanguage = NULL,
  description = NULL, localized = NULL, publishedAt = NULL,
  thumbnails = NULL, title = NULL)
}
\arguments{
\item{country}{The country of the channel}

\item{customUrl}{The custom url of the channel}

\item{defaultLanguage}{The language of the channel's default title and description}

\item{description}{The description of the channel}

\item{localized}{Localized title and description, read-only}

\item{publishedAt}{The date and time that the channel was created}

\item{thumbnails}{A map of thumbnail images associated with the channel}

\item{title}{The channel's title}
}
\value{
ChannelSnippet object
}
\description{
ChannelSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a channel, including title, description and thumbnails. Next available id: 15.
}
