% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelStatistics}
\alias{ChannelStatistics}
\title{ChannelStatistics Object}
\usage{
ChannelStatistics(commentCount = NULL, hiddenSubscriberCount = NULL,
  subscriberCount = NULL, videoCount = NULL, viewCount = NULL)
}
\arguments{
\item{commentCount}{The number of comments for the channel}

\item{hiddenSubscriberCount}{Whether or not the number of subscribers is shown for this user}

\item{subscriberCount}{The number of subscribers that the channel has}

\item{videoCount}{The number of videos uploaded to the channel}

\item{viewCount}{The number of times the channel has been viewed}
}
\value{
ChannelStatistics object
}
\description{
ChannelStatistics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Statistics about a channel: number of subscribers, number of videos in the channel, etc.
}

