% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ChannelTopicDetails}
\alias{ChannelTopicDetails}
\title{ChannelTopicDetails Object}
\usage{
ChannelTopicDetails(topicIds = NULL)
}
\arguments{
\item{topicIds}{A list of Freebase topic IDs associated with the channel}
}
\value{
ChannelTopicDetails object
}
\description{
ChannelTopicDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Freebase topic information related to the channel.
}
