% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{comments.delete}
\alias{comments.delete}
\title{Deletes a comment.}
\usage{
comments.delete(id)
}
\arguments{
\item{id}{The id parameter specifies the comment ID for the resource that is being deleted}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
