% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{comments.insert}
\alias{comments.insert}
\title{Creates a reply to an existing comment. Note: To create a top-level comment, use the commentThreads.insert method.}
\usage{
comments.insert(Comment, part)
}
\arguments{
\item{Comment}{The \link{Comment} object to pass to this method}

\item{part}{The part parameter identifies the properties that the API response will include}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other Comment functions: \code{\link{Comment}},
  \code{\link{comments.update}}
}
