% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{comments.list}
\alias{comments.list}
\title{Returns a list of comments that match the API request parameters.}
\usage{
comments.list(part, id = NULL, maxResults = NULL, pageToken = NULL,
  parentId = NULL, textFormat = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more comment resource properties that the API response will include}

\item{id}{The id parameter specifies a comma-separated list of comment IDs for the resources that are being retrieved}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}

\item{parentId}{The parentId parameter specifies the ID of the comment for which replies should be retrieved}

\item{textFormat}{This parameter indicates whether the API should return comments formatted as HTML or as plain text}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

