% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{comments.markAsSpam}
\alias{comments.markAsSpam}
\title{Expresses the caller's opinion that one or more comments should be flagged as spam.}
\usage{
comments.markAsSpam(id)
}
\arguments{
\item{id}{The id parameter specifies a comma-separated list of IDs of comments that the caller believes should be classified as spam}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

