% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{comments.setModerationStatus}
\alias{comments.setModerationStatus}
\title{Sets the moderation status of one or more comments. The API request must be authorized by the owner of the channel or video associated with the comments.}
\usage{
comments.setModerationStatus(id, moderationStatus, banAuthor = NULL)
}
\arguments{
\item{id}{The id parameter specifies a comma-separated list of IDs that identify the comments for which you are updating the moderation status}

\item{moderationStatus}{Identifies the new moderation status of the specified comments}

\item{banAuthor}{The banAuthor parameter lets you indicate that you want to automatically reject any additional comments written by the comment's author}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
