% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{CommentSnippet}
\alias{CommentSnippet}
\title{CommentSnippet Object}
\usage{
CommentSnippet(authorChannelId = NULL, authorChannelUrl = NULL,
  authorDisplayName = NULL, authorProfileImageUrl = NULL, canRate = NULL,
  channelId = NULL, likeCount = NULL, moderationStatus = NULL,
  parentId = NULL, publishedAt = NULL, textDisplay = NULL,
  textOriginal = NULL, updatedAt = NULL, videoId = NULL,
  viewerRating = NULL)
}
\arguments{
\item{authorChannelId}{The id of the author's YouTube channel, if any}

\item{authorChannelUrl}{Link to the author's YouTube channel, if any}

\item{authorDisplayName}{The name of the user who posted the comment}

\item{authorProfileImageUrl}{The URL for the avatar of the user who posted the comment}

\item{canRate}{Whether the current viewer can rate this comment}

\item{channelId}{The id of the corresponding YouTube channel}

\item{likeCount}{The total number of likes this comment has received}

\item{moderationStatus}{The comment's moderation status}

\item{parentId}{The unique id of the parent comment, only set for replies}

\item{publishedAt}{The date and time when the comment was orignally published}

\item{textDisplay}{The comment's text}

\item{textOriginal}{The comment's original raw text as initially posted or last updated}

\item{updatedAt}{The date and time when was last updated}

\item{videoId}{The ID of the video the comment refers to, if any}

\item{viewerRating}{The rating the viewer has given to this comment}
}
\value{
CommentSnippet object
}
\description{
CommentSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a comment, such as its author and text.
}
