% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{CommentThread}
\alias{CommentThread}
\title{CommentThread Object}
\usage{
CommentThread(etag = NULL, id = NULL, replies = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the comment thread}

\item{replies}{The replies object contains a limited number of replies (if any) to the top level comment found in the snippet}

\item{snippet}{The snippet object contains basic details about the comment thread and also the top level comment}
}
\value{
CommentThread object
}
\description{
CommentThread Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A comment thread represents information that applies to a top level comment and all its replies. It can also include the top level comment itself and some of the replies.
}
\seealso{
Other CommentThread functions: \code{\link{commentThreads.insert}},
  \code{\link{commentThreads.update}}
}
