% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{CommentThreadReplies}
\alias{CommentThreadReplies}
\title{CommentThreadReplies Object}
\usage{
CommentThreadReplies(comments = NULL)
}
\arguments{
\item{comments}{A limited number of replies}
}
\value{
CommentThreadReplies object
}
\description{
CommentThreadReplies Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Comments written in (direct or indirect) reply to the top level comment.
}

