% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{commentThreads.insert}
\alias{commentThreads.insert}
\title{Creates a new top-level comment. To add a reply to an existing comment, use the comments.insert method instead.}
\usage{
commentThreads.insert(CommentThread, part)
}
\arguments{
\item{CommentThread}{The \link{CommentThread} object to pass to this method}

\item{part}{The part parameter identifies the properties that the API response will include}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other CommentThread functions: \code{\link{CommentThread}},
  \code{\link{commentThreads.update}}
}
