% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{commentThreads.list}
\alias{commentThreads.list}
\title{Returns a list of comment threads that match the API request parameters.}
\usage{
commentThreads.list(part, allThreadsRelatedToChannelId = NULL,
  channelId = NULL, id = NULL, maxResults = NULL,
  moderationStatus = NULL, order = NULL, pageToken = NULL,
  searchTerms = NULL, textFormat = NULL, videoId = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more commentThread resource properties that the API response will include}

\item{allThreadsRelatedToChannelId}{The allThreadsRelatedToChannelId parameter instructs the API to return all comment threads associated with the specified channel}

\item{channelId}{The channelId parameter instructs the API to return comment threads containing comments about the specified channel}

\item{id}{The id parameter specifies a comma-separated list of comment thread IDs for the resources that should be retrieved}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{moderationStatus}{Set this parameter to limit the returned comment threads to a particular moderation state}

\item{order}{The order parameter specifies the order in which the API response should list comment threads}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}

\item{searchTerms}{The searchTerms parameter instructs the API to limit the API response to only contain comments that contain the specified search terms}

\item{textFormat}{Set this parameter's value to html or plainText to instruct the API to return the comments left by users in html formatted or in plain text}

\item{videoId}{The videoId parameter instructs the API to return comment threads associated with the specified video ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
