% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{CommentThreadSnippet}
\alias{CommentThreadSnippet}
\title{CommentThreadSnippet Object}
\usage{
CommentThreadSnippet(canReply = NULL, channelId = NULL, isPublic = NULL,
  topLevelComment = NULL, totalReplyCount = NULL, videoId = NULL)
}
\arguments{
\item{canReply}{Whether the current viewer of the thread can reply to it}

\item{channelId}{The YouTube channel the comments in the thread refer to or the channel with the video the comments refer to}

\item{isPublic}{Whether the thread (and therefore all its comments) is visible to all YouTube users}

\item{topLevelComment}{The top level comment of this thread}

\item{totalReplyCount}{The total number of replies (not including the top level comment)}

\item{videoId}{The ID of the video the comments refer to, if any}
}
\value{
CommentThreadSnippet object
}
\description{
CommentThreadSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a comment thread.
}

