% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ContentRating}
\alias{ContentRating}
\title{ContentRating Object}
\usage{
ContentRating(acbRating = NULL, agcomRating = NULL, anatelRating = NULL,
  bbfcRating = NULL, bfvcRating = NULL, bmukkRating = NULL,
  catvRating = NULL, catvfrRating = NULL, cbfcRating = NULL,
  cccRating = NULL, cceRating = NULL, chfilmRating = NULL,
  chvrsRating = NULL, cicfRating = NULL, cnaRating = NULL,
  cncRating = NULL, csaRating = NULL, cscfRating = NULL,
  czfilmRating = NULL, djctqRating = NULL, djctqRatingReasons = NULL,
  ecbmctRating = NULL, eefilmRating = NULL, egfilmRating = NULL,
  eirinRating = NULL, fcbmRating = NULL, fcoRating = NULL,
  fmocRating = NULL, fpbRating = NULL, fpbRatingReasons = NULL,
  fskRating = NULL, grfilmRating = NULL, icaaRating = NULL,
  ifcoRating = NULL, ilfilmRating = NULL, incaaRating = NULL,
  kfcbRating = NULL, kijkwijzerRating = NULL, kmrbRating = NULL,
  lsfRating = NULL, mccaaRating = NULL, mccypRating = NULL,
  mdaRating = NULL, medietilsynetRating = NULL, mekuRating = NULL,
  mibacRating = NULL, mocRating = NULL, moctwRating = NULL,
  mpaaRating = NULL, mtrcbRating = NULL, nbcRating = NULL,
  nbcplRating = NULL, nfrcRating = NULL, nfvcbRating = NULL,
  nkclvRating = NULL, oflcRating = NULL, pefilmRating = NULL,
  rcnofRating = NULL, resorteviolenciaRating = NULL, rtcRating = NULL,
  rteRating = NULL, russiaRating = NULL, skfilmRating = NULL,
  smaisRating = NULL, smsaRating = NULL, tvpgRating = NULL,
  ytRating = NULL)
}
\arguments{
\item{acbRating}{The video's Australian Classification Board (ACB) or Australian Communications and Media Authority (ACMA) rating}

\item{agcomRating}{The video's rating from Italy's Autorità per le Garanzie nelle Comunicazioni (AGCOM)}

\item{anatelRating}{The video's Anatel (Asociación Nacional de Televisión) rating for Chilean television}

\item{bbfcRating}{The video's British Board of Film Classification (BBFC) rating}

\item{bfvcRating}{The video's rating from Thailand's Board of Film and Video Censors}

\item{bmukkRating}{The video's rating from the Austrian Board of Media Classification (Bundesministerium für Unterricht, Kunst und Kultur)}

\item{catvRating}{Rating system for Canadian TV - Canadian TV Classification System The video's rating from the Canadian Radio-Television and Telecommunications Commission (CRTC) for Canadian English-language broadcasts}

\item{catvfrRating}{The video's rating from the Canadian Radio-Television and Telecommunications Commission (CRTC) for Canadian French-language broadcasts}

\item{cbfcRating}{The video's Central Board of Film Certification (CBFC - India) rating}

\item{cccRating}{The video's Consejo de Calificación Cinematográfica (Chile) rating}

\item{cceRating}{The video's rating from Portugal's Comissão de Classificação de Espect´culos}

\item{chfilmRating}{The video's rating in Switzerland}

\item{chvrsRating}{The video's Canadian Home Video Rating System (CHVRS) rating}

\item{cicfRating}{The video's rating from the Commission de Contrôle des Films (Belgium)}

\item{cnaRating}{The video's rating from Romania's CONSILIUL NATIONAL AL AUDIOVIZUALULUI (CNA)}

\item{cncRating}{Rating system in France - Commission de classification cinematographique}

\item{csaRating}{The video's rating from France's Conseil supérieur de l?audiovisuel, which rates broadcast content}

\item{cscfRating}{The video's rating from Luxembourg's Commission de surveillance de la classification des films (CSCF)}

\item{czfilmRating}{The video's rating in the Czech Republic}

\item{djctqRating}{The video's Departamento de Justiça, Classificação, Qualificação e Títulos (DJCQT - Brazil) rating}

\item{djctqRatingReasons}{Reasons that explain why the video received its DJCQT (Brazil) rating}

\item{ecbmctRating}{Rating system in Turkey - Evaluation and Classification Board of the Ministry of Culture and Tourism}

\item{eefilmRating}{The video's rating in Estonia}

\item{egfilmRating}{The video's rating in Egypt}

\item{eirinRating}{The video's Eirin (映倫) rating}

\item{fcbmRating}{The video's rating from Malaysia's Film Censorship Board}

\item{fcoRating}{The video's rating from Hong Kong's Office for Film, Newspaper and Article Administration}

\item{fmocRating}{This property has been deprecated}

\item{fpbRating}{The video's rating from South Africa's Film and Publication Board}

\item{fpbRatingReasons}{Reasons that explain why the video received its FPB (South Africa) rating}

\item{fskRating}{The video's Freiwillige Selbstkontrolle der Filmwirtschaft (FSK - Germany) rating}

\item{grfilmRating}{The video's rating in Greece}

\item{icaaRating}{The video's Instituto de la Cinematografía y de las Artes Audiovisuales (ICAA - Spain) rating}

\item{ifcoRating}{The video's Irish Film Classification Office (IFCO - Ireland) rating}

\item{ilfilmRating}{The video's rating in Israel}

\item{incaaRating}{The video's INCAA (Instituto Nacional de Cine y Artes Audiovisuales - Argentina) rating}

\item{kfcbRating}{The video's rating from the Kenya Film Classification Board}

\item{kijkwijzerRating}{voor de Classificatie van Audiovisuele Media (Netherlands)}

\item{kmrbRating}{The video's Korea Media Rating Board (영상물등급위원회) rating}

\item{lsfRating}{The video's rating from Indonesia's Lembaga Sensor Film}

\item{mccaaRating}{The video's rating from Malta's Film Age-Classification Board}

\item{mccypRating}{The video's rating from the Danish Film Institute's (Det Danske Filminstitut) Media Council for Children and Young People}

\item{mdaRating}{The video's rating from Singapore's Media Development Authority (MDA) and, specifically, it's Board of Film Censors (BFC)}

\item{medietilsynetRating}{The video's rating from Medietilsynet, the Norwegian Media Authority}

\item{mekuRating}{The video's rating from Finland's Kansallinen Audiovisuaalinen Instituutti (National Audiovisual Institute)}

\item{mibacRating}{The video's rating from the Ministero dei Beni e delle Attività Culturali e del Turismo (Italy)}

\item{mocRating}{The video's Ministerio de Cultura (Colombia) rating}

\item{moctwRating}{The video's rating from Taiwan's Ministry of Culture (文化部)}

\item{mpaaRating}{The video's Motion Picture Association of America (MPAA) rating}

\item{mtrcbRating}{The video's rating from the Movie and Television Review and Classification Board (Philippines)}

\item{nbcRating}{The video's rating from the Maldives National Bureau of Classification}

\item{nbcplRating}{The video's rating in Poland}

\item{nfrcRating}{The video's rating from the Bulgarian National Film Center}

\item{nfvcbRating}{The video's rating from Nigeria's National Film and Video Censors Board}

\item{nkclvRating}{The video's rating from the Nacionãlais Kino centrs (National Film Centre of Latvia)}

\item{oflcRating}{The video's Office of Film and Literature Classification (OFLC - New Zealand) rating}

\item{pefilmRating}{The video's rating in Peru}

\item{rcnofRating}{The video's rating from the Hungarian Nemzeti Filmiroda, the Rating Committee of the National Office of Film}

\item{resorteviolenciaRating}{The video's rating in Venezuela}

\item{rtcRating}{The video's General Directorate of Radio, Television and Cinematography (Mexico) rating}

\item{rteRating}{The video's rating from Ireland's Raidió Teilifís Éireann}

\item{russiaRating}{The video's National Film Registry of the Russian Federation (MKRF - Russia) rating}

\item{skfilmRating}{The video's rating in Slovakia}

\item{smaisRating}{The video's rating in Iceland}

\item{smsaRating}{The video's rating from Statens medieråd (Sweden's National Media Council)}

\item{tvpgRating}{The video's TV Parental Guidelines (TVPG) rating}

\item{ytRating}{A rating that YouTube uses to identify age-restricted content}
}
\value{
ContentRating object
}
\description{
ContentRating Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Ratings schemes. The country-specific ratings are mostly for movies and shows. NEXT_ID: 68
}
