% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{FanFundingEvent}
\alias{FanFundingEvent}
\title{FanFundingEvent Object}
\usage{
FanFundingEvent(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the fan funding event}

\item{snippet}{The snippet object contains basic details about the fan funding event}
}
\value{
FanFundingEvent object
}
\description{
FanFundingEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A fanFundingEvent resource represents a fan funding event on a YouTube channel. Fan funding events occur when a user gives one-time monetary support to the channel owner.
}

