% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{FanFundingEventSnippet}
\alias{FanFundingEventSnippet}
\title{FanFundingEventSnippet Object}
\usage{
FanFundingEventSnippet(amountMicros = NULL, channelId = NULL,
  commentText = NULL, createdAt = NULL, currency = NULL,
  displayString = NULL, supporterDetails = NULL)
}
\arguments{
\item{amountMicros}{The amount of funding in micros of fund_currency}

\item{channelId}{Channel id where the funding event occurred}

\item{commentText}{The text contents of the comment left by the user}

\item{createdAt}{The date and time when the funding occurred}

\item{currency}{The currency in which the fund was made}

\item{displayString}{A rendered string that displays the fund amount and currency (e}

\item{supporterDetails}{Details about the supporter}
}
\value{
FanFundingEventSnippet object
}
\description{
FanFundingEventSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
