% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{GeoPoint}
\alias{GeoPoint}
\title{GeoPoint Object}
\usage{
GeoPoint(altitude = NULL, latitude = NULL, longitude = NULL)
}
\arguments{
\item{altitude}{Altitude above the reference ellipsoid, in meters}

\item{latitude}{Latitude in degrees}

\item{longitude}{Longitude in degrees}
}
\value{
GeoPoint object
}
\description{
GeoPoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Geographical coordinates of a point, in WGS84.
}

