% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{GuideCategory}
\alias{GuideCategory}
\title{GuideCategory Object}
\usage{
GuideCategory(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the guide category}

\item{snippet}{The snippet object contains basic details about the category, such as its title}
}
\value{
GuideCategory object
}
\description{
GuideCategory Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A guideCategory resource identifies a category that YouTube algorithmically assigns based on a channel's content or other indicators, such as the channel's popularity. The list is similar to video categories, with the difference being that a video's uploader can assign a video category but only YouTube can assign a channel category.
}

