% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{GuideCategorySnippet}
\alias{GuideCategorySnippet}
\title{GuideCategorySnippet Object}
\usage{
GuideCategorySnippet(channelId = NULL, title = NULL)
}
\arguments{
\item{channelId}{No description}

\item{title}{Description of the guide category}
}
\value{
GuideCategorySnippet object
}
\description{
GuideCategorySnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a guide category.
}
