% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{I18nLanguage}
\alias{I18nLanguage}
\title{I18nLanguage Object}
\usage{
I18nLanguage(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the i18n language}

\item{snippet}{The snippet object contains basic details about the i18n language, such as language code and human-readable name}
}
\value{
I18nLanguage object
}
\description{
I18nLanguage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An i18nLanguage resource identifies a UI language currently supported by YouTube.
}

