% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{I18nLanguageListResponse}
\alias{I18nLanguageListResponse}
\title{I18nLanguageListResponse Object}
\usage{
I18nLanguageListResponse(etag = NULL, eventId = NULL, items = NULL,
  visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of supported i18n languages}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
I18nLanguageListResponse object
}
\description{
I18nLanguageListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

