% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{I18nLanguageSnippet}
\alias{I18nLanguageSnippet}
\title{I18nLanguageSnippet Object}
\usage{
I18nLanguageSnippet(hl = NULL, name = NULL)
}
\arguments{
\item{hl}{A short BCP-47 code that uniquely identifies a language}

\item{name}{The human-readable name of the language in the language itself}
}
\value{
I18nLanguageSnippet object
}
\description{
I18nLanguageSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about an i18n language, such as language code and human-readable name.
}
