% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{I18nRegion}
\alias{I18nRegion}
\title{I18nRegion Object}
\usage{
I18nRegion(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the i18n region}

\item{snippet}{The snippet object contains basic details about the i18n region, such as region code and human-readable name}
}
\value{
I18nRegion object
}
\description{
I18nRegion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A i18nRegion resource identifies a region where YouTube is available.
}

