% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{I18nRegionListResponse}
\alias{I18nRegionListResponse}
\title{I18nRegionListResponse Object}
\usage{
I18nRegionListResponse(etag = NULL, eventId = NULL, items = NULL,
  visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of regions where YouTube is available}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
I18nRegionListResponse object
}
\description{
I18nRegionListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

