% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{I18nRegionSnippet}
\alias{I18nRegionSnippet}
\title{I18nRegionSnippet Object}
\usage{
I18nRegionSnippet(gl = NULL, name = NULL)
}
\arguments{
\item{gl}{The region code as a 2-letter ISO country code}

\item{name}{The human-readable name of the region}
}
\value{
I18nRegionSnippet object
}
\description{
I18nRegionSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about an i18n region, such as region code and human-readable name.
}

