% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ImageSettings}
\alias{ImageSettings}
\title{ImageSettings Object}
\usage{
ImageSettings(backgroundImageUrl = NULL, bannerExternalUrl = NULL,
  bannerImageUrl = NULL, bannerMobileExtraHdImageUrl = NULL,
  bannerMobileHdImageUrl = NULL, bannerMobileImageUrl = NULL,
  bannerMobileLowImageUrl = NULL, bannerMobileMediumHdImageUrl = NULL,
  bannerTabletExtraHdImageUrl = NULL, bannerTabletHdImageUrl = NULL,
  bannerTabletImageUrl = NULL, bannerTabletLowImageUrl = NULL,
  bannerTvHighImageUrl = NULL, bannerTvImageUrl = NULL,
  bannerTvLowImageUrl = NULL, bannerTvMediumImageUrl = NULL,
  largeBrandedBannerImageImapScript = NULL,
  largeBrandedBannerImageUrl = NULL,
  smallBrandedBannerImageImapScript = NULL,
  smallBrandedBannerImageUrl = NULL, trackingImageUrl = NULL,
  watchIconImageUrl = NULL)
}
\arguments{
\item{backgroundImageUrl}{The URL for the background image shown on the video watch page}

\item{bannerExternalUrl}{This is used only in update requests; if it's set, we use this URL to generate all of the above banner URLs}

\item{bannerImageUrl}{Banner image}

\item{bannerMobileExtraHdImageUrl}{Banner image}

\item{bannerMobileHdImageUrl}{Banner image}

\item{bannerMobileImageUrl}{Banner image}

\item{bannerMobileLowImageUrl}{Banner image}

\item{bannerMobileMediumHdImageUrl}{Banner image}

\item{bannerTabletExtraHdImageUrl}{Banner image}

\item{bannerTabletHdImageUrl}{Banner image}

\item{bannerTabletImageUrl}{Banner image}

\item{bannerTabletLowImageUrl}{Banner image}

\item{bannerTvHighImageUrl}{Banner image}

\item{bannerTvImageUrl}{Banner image}

\item{bannerTvLowImageUrl}{Banner image}

\item{bannerTvMediumImageUrl}{Banner image}

\item{largeBrandedBannerImageImapScript}{The image map script for the large banner image}

\item{largeBrandedBannerImageUrl}{The URL for the 854px by 70px image that appears below the video player in the expanded video view of the video watch page}

\item{smallBrandedBannerImageImapScript}{The image map script for the small banner image}

\item{smallBrandedBannerImageUrl}{The URL for the 640px by 70px banner image that appears below the video player in the default view of the video watch page}

\item{trackingImageUrl}{The URL for a 1px by 1px tracking pixel that can be used to collect statistics for views of the channel or video pages}

\item{watchIconImageUrl}{The URL for the image that appears above the top-left corner of the video player}
}
\value{
ImageSettings object
}
\description{
ImageSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Branding properties for images associated with the channel.
}

