% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{IngestionInfo}
\alias{IngestionInfo}
\title{IngestionInfo Object}
\usage{
IngestionInfo(backupIngestionAddress = NULL, ingestionAddress = NULL,
  streamName = NULL)
}
\arguments{
\item{backupIngestionAddress}{The backup ingestion URL that you should use to stream video to YouTube}

\item{ingestionAddress}{The primary ingestion URL that you should use to stream video to YouTube}

\item{streamName}{The HTTP or RTMP stream name that YouTube assigns to the video stream}
}
\value{
IngestionInfo object
}
\description{
IngestionInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes information necessary for ingesting an RTMP or an HTTP stream.
}

