% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{InvideoBranding}
\alias{InvideoBranding}
\title{InvideoBranding Object}
\usage{
InvideoBranding(imageBytes = NULL, imageUrl = NULL, position = NULL,
  targetChannelId = NULL, timing = NULL)
}
\arguments{
\item{imageBytes}{No description}

\item{imageUrl}{No description}

\item{position}{No description}

\item{targetChannelId}{No description}

\item{timing}{No description}
}
\value{
InvideoBranding object
}
\description{
InvideoBranding Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InvideoBranding functions: \code{\link{watermarks.set}}
}
