% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{InvideoPosition}
\alias{InvideoPosition}
\title{InvideoPosition Object}
\usage{
InvideoPosition(cornerPosition = NULL, type = NULL)
}
\arguments{
\item{cornerPosition}{Describes in which corner of the video the visual widget will appear}

\item{type}{Defines the position type}
}
\value{
InvideoPosition object
}
\description{
InvideoPosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the spatial position of a visual widget inside a video. It is a union of various position types, out of which only will be set one.
}

