% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{InvideoPromotion}
\alias{InvideoPromotion}
\title{InvideoPromotion Object}
\usage{
InvideoPromotion(defaultTiming = NULL, items = NULL, position = NULL,
  useSmartTiming = NULL)
}
\arguments{
\item{defaultTiming}{The default temporal position within the video where the promoted item will be displayed}

\item{items}{List of promoted items in decreasing priority}

\item{position}{The spatial position within the video where the promoted item will be displayed}

\item{useSmartTiming}{Indicates whether the channel's promotional campaign uses 'smart timing}
}
\value{
InvideoPromotion object
}
\description{
InvideoPromotion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes an invideo promotion campaign consisting of multiple promoted items. A campaign belongs to a single channel_id.
}

