% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{InvideoTiming}
\alias{InvideoTiming}
\title{InvideoTiming Object}
\usage{
InvideoTiming(durationMs = NULL, offsetMs = NULL, type = NULL)
}
\arguments{
\item{durationMs}{Defines the duration in milliseconds for which the promotion should be displayed}

\item{offsetMs}{Defines the time at which the promotion will appear}

\item{type}{Describes a timing type}
}
\value{
InvideoTiming object
}
\description{
InvideoTiming Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a temporal position of a visual widget inside a video.
}

