% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LanguageTag}
\alias{LanguageTag}
\title{LanguageTag Object}
\usage{
LanguageTag(value = NULL)
}
\arguments{
\item{value}{No description}
}
\value{
LanguageTag object
}
\description{
LanguageTag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

