% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcast}
\alias{LiveBroadcast}
\title{LiveBroadcast Object}
\usage{
LiveBroadcast(contentDetails = NULL, etag = NULL, id = NULL,
  snippet = NULL, statistics = NULL, status = NULL, topicDetails = NULL)
}
\arguments{
\item{contentDetails}{The contentDetails object contains information about the event's video content, such as whether the content can be shown in an embedded video player or if it will be archived and therefore available for viewing after the event has concluded}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the broadcast}

\item{snippet}{The snippet object contains basic details about the event, including its title, description, start time, and end time}

\item{statistics}{The statistics object contains info about the event's current stats}

\item{status}{The status object contains information about the event's status}

\item{topicDetails}{No description}
}
\value{
LiveBroadcast object
}
\description{
LiveBroadcast Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A liveBroadcast resource represents an event that will be streamed, via live video, on YouTube.
}
\seealso{
Other LiveBroadcast functions: \code{\link{liveBroadcasts.insert}},
  \code{\link{liveBroadcasts.update}}
}

