% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcastContentDetails}
\alias{LiveBroadcastContentDetails}
\title{LiveBroadcastContentDetails Object}
\usage{
LiveBroadcastContentDetails(boundStreamId = NULL,
  boundStreamLastUpdateTimeMs = NULL, closedCaptionsType = NULL,
  enableClosedCaptions = NULL, enableContentEncryption = NULL,
  enableDvr = NULL, enableEmbed = NULL, enableLowLatency = NULL,
  monitorStream = NULL, projection = NULL, recordFromStart = NULL,
  startWithSlate = NULL)
}
\arguments{
\item{boundStreamId}{This value uniquely identifies the live stream bound to the broadcast}

\item{boundStreamLastUpdateTimeMs}{The date and time that the live stream referenced by boundStreamId was last updated}

\item{closedCaptionsType}{No description}

\item{enableClosedCaptions}{This setting indicates whether HTTP POST closed captioning is enabled for this broadcast}

\item{enableContentEncryption}{This setting indicates whether YouTube should enable content encryption for the broadcast}

\item{enableDvr}{This setting determines whether viewers can access DVR controls while watching the video}

\item{enableEmbed}{This setting indicates whether the broadcast video can be played in an embedded player}

\item{enableLowLatency}{Indicates whether this broadcast has low latency enabled}

\item{monitorStream}{The monitorStream object contains information about the monitor stream, which the broadcaster can use to review the event content before the broadcast stream is shown publicly}

\item{projection}{The projection format of this broadcast}

\item{recordFromStart}{Automatically start recording after the event goes live}

\item{startWithSlate}{This setting indicates whether the broadcast should automatically begin with an in-stream slate when you update the broadcast's status to live}
}
\value{
LiveBroadcastContentDetails object
}
\description{
LiveBroadcastContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detailed settings of a broadcast.
}

