% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{liveBroadcasts.bind}
\alias{liveBroadcasts.bind}
\title{Binds a YouTube broadcast to a stream or removes an existing binding between a broadcast and a stream. A broadcast can only be bound to one video stream, though a video stream may be bound to more than one broadcast.}
\usage{
liveBroadcasts.bind(id, part, onBehalfOfContentOwner = NULL,
  onBehalfOfContentOwnerChannel = NULL, streamId = NULL)
}
\arguments{
\item{id}{The id parameter specifies the unique ID of the broadcast that is being bound to a video stream}

\item{part}{The part parameter specifies a comma-separated list of one or more liveBroadcast resource properties that the API response will include}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}

\item{streamId}{The streamId parameter specifies the unique ID of the video stream that is being bound to a broadcast}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
