% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{liveBroadcasts.control}
\alias{liveBroadcasts.control}
\title{Controls the settings for a slate that can be displayed in the broadcast stream.}
\usage{
liveBroadcasts.control(id, part, displaySlate = NULL, offsetTimeMs = NULL,
  onBehalfOfContentOwner = NULL, onBehalfOfContentOwnerChannel = NULL,
  walltime = NULL)
}
\arguments{
\item{id}{The id parameter specifies the YouTube live broadcast ID that uniquely identifies the broadcast in which the slate is being updated}

\item{part}{The part parameter specifies a comma-separated list of one or more liveBroadcast resource properties that the API response will include}

\item{displaySlate}{The displaySlate parameter specifies whether the slate is being enabled or disabled}

\item{offsetTimeMs}{The offsetTimeMs parameter specifies a positive time offset when the specified slate change will occur}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}

\item{walltime}{The walltime parameter specifies the wall clock time at which the specified slate change will occur}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
