% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{liveBroadcasts.list}
\alias{liveBroadcasts.list}
\title{Returns a list of YouTube broadcasts that match the API request parameters.}
\usage{
liveBroadcasts.list(part, broadcastStatus = NULL, broadcastType = NULL,
  id = NULL, maxResults = NULL, mine = NULL,
  onBehalfOfContentOwner = NULL, onBehalfOfContentOwnerChannel = NULL,
  pageToken = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more liveBroadcast resource properties that the API response will include}

\item{broadcastStatus}{The broadcastStatus parameter filters the API response to only include broadcasts with the specified status}

\item{broadcastType}{The broadcastType parameter filters the API response to only include broadcasts with the specified type}

\item{id}{The id parameter specifies a comma-separated list of YouTube broadcast IDs that identify the broadcasts being retrieved}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{mine}{The mine parameter can be used to instruct the API to only return broadcasts owned by the authenticated user}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

