% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{liveBroadcasts.update}
\alias{liveBroadcasts.update}
\title{Updates a broadcast. For example, you could modify the broadcast settings defined in the liveBroadcast resource's contentDetails object.}
\usage{
liveBroadcasts.update(LiveBroadcast, part, onBehalfOfContentOwner = NULL,
  onBehalfOfContentOwnerChannel = NULL)
}
\arguments{
\item{LiveBroadcast}{The \link{LiveBroadcast} object to pass to this method}

\item{part}{The part parameter serves two purposes in this operation}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other LiveBroadcast functions: \code{\link{LiveBroadcast}},
  \code{\link{liveBroadcasts.insert}}
}

