% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcastSnippet}
\alias{LiveBroadcastSnippet}
\title{LiveBroadcastSnippet Object}
\usage{
LiveBroadcastSnippet(actualEndTime = NULL, actualStartTime = NULL,
  channelId = NULL, description = NULL, isDefaultBroadcast = NULL,
  liveChatId = NULL, publishedAt = NULL, scheduledEndTime = NULL,
  scheduledStartTime = NULL, thumbnails = NULL, title = NULL)
}
\arguments{
\item{actualEndTime}{The date and time that the broadcast actually ended}

\item{actualStartTime}{The date and time that the broadcast actually started}

\item{channelId}{The ID that YouTube uses to uniquely identify the channel that is publishing the broadcast}

\item{description}{The broadcast's description}

\item{isDefaultBroadcast}{No description}

\item{liveChatId}{The id of the live chat for this broadcast}

\item{publishedAt}{The date and time that the broadcast was added to YouTube's live broadcast schedule}

\item{scheduledEndTime}{The date and time that the broadcast is scheduled to end}

\item{scheduledStartTime}{The date and time that the broadcast is scheduled to start}

\item{thumbnails}{A map of thumbnail images associated with the broadcast}

\item{title}{The broadcast's title}
}
\value{
LiveBroadcastSnippet object
}
\description{
LiveBroadcastSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
