% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcastStatistics}
\alias{LiveBroadcastStatistics}
\title{LiveBroadcastStatistics Object}
\usage{
LiveBroadcastStatistics(concurrentViewers = NULL, totalChatCount = NULL)
}
\arguments{
\item{concurrentViewers}{The number of viewers currently watching the broadcast}

\item{totalChatCount}{The total number of live chat messages currently on the broadcast}
}
\value{
LiveBroadcastStatistics object
}
\description{
LiveBroadcastStatistics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Statistics about the live broadcast. These represent a snapshot of the values at the time of the request. Statistics are only returned for live broadcasts.
}
