% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcastStatus}
\alias{LiveBroadcastStatus}
\title{LiveBroadcastStatus Object}
\usage{
LiveBroadcastStatus(lifeCycleStatus = NULL, liveBroadcastPriority = NULL,
  privacyStatus = NULL, recordingStatus = NULL)
}
\arguments{
\item{lifeCycleStatus}{The broadcast's status}

\item{liveBroadcastPriority}{Priority of the live broadcast event (internal state)}

\item{privacyStatus}{The broadcast's privacy status}

\item{recordingStatus}{The broadcast's recording status}
}
\value{
LiveBroadcastStatus object
}
\description{
LiveBroadcastStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
