% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcastTopic}
\alias{LiveBroadcastTopic}
\title{LiveBroadcastTopic Object}
\usage{
LiveBroadcastTopic(snippet = NULL, type = NULL, unmatched = NULL)
}
\arguments{
\item{snippet}{Information about the topic matched}

\item{type}{The type of the topic}

\item{unmatched}{If this flag is set it means that we have not been able to match the topic title and type provided to a known entity}
}
\value{
LiveBroadcastTopic object
}
\description{
LiveBroadcastTopic Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
