% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcastTopicDetails}
\alias{LiveBroadcastTopicDetails}
\title{LiveBroadcastTopicDetails Object}
\usage{
LiveBroadcastTopicDetails(topics = NULL)
}
\arguments{
\item{topics}{No description}
}
\value{
LiveBroadcastTopicDetails object
}
\description{
LiveBroadcastTopicDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

