% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveBroadcastTopicSnippet}
\alias{LiveBroadcastTopicSnippet}
\title{LiveBroadcastTopicSnippet Object}
\usage{
LiveBroadcastTopicSnippet(name = NULL, releaseDate = NULL)
}
\arguments{
\item{name}{The name of the topic}

\item{releaseDate}{The date at which the topic was released}
}
\value{
LiveBroadcastTopicSnippet object
}
\description{
LiveBroadcastTopicSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
