% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatBan}
\alias{LiveChatBan}
\title{LiveChatBan Object}
\usage{
LiveChatBan(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the ban}

\item{snippet}{The snippet object contains basic details about the ban}
}
\value{
LiveChatBan object
}
\description{
LiveChatBan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A liveChatBan resource represents a ban for a YouTube live chat.
}
\seealso{
Other LiveChatBan functions: \code{\link{liveChatBans.insert}}
}
