% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{liveChatBans.insert}
\alias{liveChatBans.insert}
\title{Adds a new ban to the chat.}
\usage{
liveChatBans.insert(LiveChatBan, part)
}
\arguments{
\item{LiveChatBan}{The \link{LiveChatBan} object to pass to this method}

\item{part}{The part parameter serves two purposes in this operation}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other LiveChatBan functions: \code{\link{LiveChatBan}}
}
