% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatBanSnippet}
\alias{LiveChatBanSnippet}
\title{LiveChatBanSnippet Object}
\usage{
LiveChatBanSnippet(banDurationSeconds = NULL, bannedUserDetails = NULL,
  liveChatId = NULL, type = NULL)
}
\arguments{
\item{banDurationSeconds}{The duration of a ban, only filled if the ban has type TEMPORARY}

\item{bannedUserDetails}{No description}

\item{liveChatId}{The chat this ban is pertinent to}

\item{type}{The type of ban}
}
\value{
LiveChatBanSnippet object
}
\description{
LiveChatBanSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
