% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatFanFundingEventDetails}
\alias{LiveChatFanFundingEventDetails}
\title{LiveChatFanFundingEventDetails Object}
\usage{
LiveChatFanFundingEventDetails(amountDisplayString = NULL,
  amountMicros = NULL, currency = NULL, userComment = NULL)
}
\arguments{
\item{amountDisplayString}{A rendered string that displays the fund amount and currency to the user}

\item{amountMicros}{The amount of the fund}

\item{currency}{The currency in which the fund was made}

\item{userComment}{The comment added by the user to this fan funding event}
}
\value{
LiveChatFanFundingEventDetails object
}
\description{
LiveChatFanFundingEventDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
