% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatMessage}
\alias{LiveChatMessage}
\title{LiveChatMessage Object}
\usage{
LiveChatMessage(authorDetails = NULL, etag = NULL, id = NULL,
  snippet = NULL)
}
\arguments{
\item{authorDetails}{The authorDetails object contains basic details about the user that posted this message}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the message}

\item{snippet}{The snippet object contains basic details about the message}
}
\value{
LiveChatMessage object
}
\description{
LiveChatMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A liveChatMessage resource represents a chat message in a YouTube Live Chat.
}
\seealso{
Other LiveChatMessage functions: \code{\link{liveChatMessages.insert}}
}
