% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatMessageAuthorDetails}
\alias{LiveChatMessageAuthorDetails}
\title{LiveChatMessageAuthorDetails Object}
\usage{
LiveChatMessageAuthorDetails(channelId = NULL, channelUrl = NULL,
  displayName = NULL, isChatModerator = NULL, isChatOwner = NULL,
  isChatSponsor = NULL, isVerified = NULL, profileImageUrl = NULL)
}
\arguments{
\item{channelId}{The YouTube channel ID}

\item{channelUrl}{The channel's URL}

\item{displayName}{The channel's display name}

\item{isChatModerator}{Whether the author is a moderator of the live chat}

\item{isChatOwner}{Whether the author is the owner of the live chat}

\item{isChatSponsor}{Whether the author is a sponsor of the live chat}

\item{isVerified}{Whether the author's identity has been verified by YouTube}

\item{profileImageUrl}{The channels's avatar URL}
}
\value{
LiveChatMessageAuthorDetails object
}
\description{
LiveChatMessageAuthorDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

