% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatMessageDeletedDetails}
\alias{LiveChatMessageDeletedDetails}
\title{LiveChatMessageDeletedDetails Object}
\usage{
LiveChatMessageDeletedDetails(deletedMessageId = NULL)
}
\arguments{
\item{deletedMessageId}{No description}
}
\value{
LiveChatMessageDeletedDetails object
}
\description{
LiveChatMessageDeletedDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
