% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatMessageListResponse}
\alias{LiveChatMessageListResponse}
\title{LiveChatMessageListResponse Object}
\usage{
LiveChatMessageListResponse(etag = NULL, eventId = NULL, items = NULL,
  nextPageToken = NULL, offlineAt = NULL, pageInfo = NULL,
  pollingIntervalMillis = NULL, tokenPagination = NULL, visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of live chat messages}

\item{nextPageToken}{The token that can be used as the value of the pageToken parameter to retrieve the next page in the result set}

\item{offlineAt}{The date and time when the underlying stream went offline}

\item{pageInfo}{No description}

\item{pollingIntervalMillis}{The amount of time the client should wait before polling again}

\item{tokenPagination}{No description}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
LiveChatMessageListResponse object
}
\description{
LiveChatMessageListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
