% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatMessageRetractedDetails}
\alias{LiveChatMessageRetractedDetails}
\title{LiveChatMessageRetractedDetails Object}
\usage{
LiveChatMessageRetractedDetails(retractedMessageId = NULL)
}
\arguments{
\item{retractedMessageId}{No description}
}
\value{
LiveChatMessageRetractedDetails object
}
\description{
LiveChatMessageRetractedDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
