% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{liveChatMessages.insert}
\alias{liveChatMessages.insert}
\title{Adds a message to a live chat.}
\usage{
liveChatMessages.insert(LiveChatMessage, part)
}
\arguments{
\item{LiveChatMessage}{The \link{LiveChatMessage} object to pass to this method}

\item{part}{The part parameter serves two purposes}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other LiveChatMessage functions: \code{\link{LiveChatMessage}}
}

