% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatMessageSnippet}
\alias{LiveChatMessageSnippet}
\title{LiveChatMessageSnippet Object}
\usage{
LiveChatMessageSnippet(authorChannelId = NULL, displayMessage = NULL,
  fanFundingEventDetails = NULL, hasDisplayContent = NULL,
  liveChatId = NULL, messageDeletedDetails = NULL,
  messageRetractedDetails = NULL, pollClosedDetails = NULL,
  pollEditedDetails = NULL, pollOpenedDetails = NULL,
  pollVotedDetails = NULL, publishedAt = NULL, textMessageDetails = NULL,
  type = NULL, userBannedDetails = NULL)
}
\arguments{
\item{authorChannelId}{The ID of the user that authored this message, this field is not always filled}

\item{displayMessage}{Contains a string that can be displayed to the user}

\item{fanFundingEventDetails}{Details about the funding event, this is only set if the type is 'fanFundingEvent'}

\item{hasDisplayContent}{Whether the message has display content that should be displayed to users}

\item{liveChatId}{No description}

\item{messageDeletedDetails}{No description}

\item{messageRetractedDetails}{No description}

\item{pollClosedDetails}{No description}

\item{pollEditedDetails}{No description}

\item{pollOpenedDetails}{No description}

\item{pollVotedDetails}{No description}

\item{publishedAt}{The date and time when the message was orignally published}

\item{textMessageDetails}{Details about the text message, this is only set if the type is 'textMessageEvent'}

\item{type}{The type of message, this will always be present, it determines the contents of the message as well as which fields will be present}

\item{userBannedDetails}{No description}
}
\value{
LiveChatMessageSnippet object
}
\description{
LiveChatMessageSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

