% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatModerator}
\alias{LiveChatModerator}
\title{LiveChatModerator Object}
\usage{
LiveChatModerator(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the moderator}

\item{snippet}{The snippet object contains basic details about the moderator}
}
\value{
LiveChatModerator object
}
\description{
LiveChatModerator Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A liveChatModerator resource represents a moderator for a YouTube live chat. A chat moderator has the ability to ban/unban users from a chat, remove message, etc.
}
\seealso{
Other LiveChatModerator functions: \code{\link{liveChatModerators.insert}}
}
