% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatModeratorSnippet}
\alias{LiveChatModeratorSnippet}
\title{LiveChatModeratorSnippet Object}
\usage{
LiveChatModeratorSnippet(liveChatId = NULL, moderatorDetails = NULL)
}
\arguments{
\item{liveChatId}{The ID of the live chat this moderator can act on}

\item{moderatorDetails}{Details about the moderator}
}
\value{
LiveChatModeratorSnippet object
}
\description{
LiveChatModeratorSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

