% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatPollClosedDetails}
\alias{LiveChatPollClosedDetails}
\title{LiveChatPollClosedDetails Object}
\usage{
LiveChatPollClosedDetails(pollId = NULL)
}
\arguments{
\item{pollId}{The id of the poll that was closed}
}
\value{
LiveChatPollClosedDetails object
}
\description{
LiveChatPollClosedDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
