% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatPollEditedDetails}
\alias{LiveChatPollEditedDetails}
\title{LiveChatPollEditedDetails Object}
\usage{
LiveChatPollEditedDetails(id = NULL, items = NULL, prompt = NULL)
}
\arguments{
\item{id}{No description}

\item{items}{No description}

\item{prompt}{No description}
}
\value{
LiveChatPollEditedDetails object
}
\description{
LiveChatPollEditedDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
