% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatPollItem}
\alias{LiveChatPollItem}
\title{LiveChatPollItem Object}
\usage{
LiveChatPollItem(description = NULL, itemId = NULL)
}
\arguments{
\item{description}{Plain text description of the item}

\item{itemId}{No description}
}
\value{
LiveChatPollItem object
}
\description{
LiveChatPollItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
