% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatPollOpenedDetails}
\alias{LiveChatPollOpenedDetails}
\title{LiveChatPollOpenedDetails Object}
\usage{
LiveChatPollOpenedDetails(id = NULL, items = NULL, prompt = NULL)
}
\arguments{
\item{id}{No description}

\item{items}{No description}

\item{prompt}{No description}
}
\value{
LiveChatPollOpenedDetails object
}
\description{
LiveChatPollOpenedDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

