% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatPollVotedDetails}
\alias{LiveChatPollVotedDetails}
\title{LiveChatPollVotedDetails Object}
\usage{
LiveChatPollVotedDetails(itemId = NULL, pollId = NULL)
}
\arguments{
\item{itemId}{The poll item the user chose}

\item{pollId}{The poll the user voted on}
}
\value{
LiveChatPollVotedDetails object
}
\description{
LiveChatPollVotedDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
