% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatSuperChatDetails}
\alias{LiveChatSuperChatDetails}
\title{LiveChatSuperChatDetails Object}
\usage{
LiveChatSuperChatDetails(amountDisplayString = NULL, amountMicros = NULL,
  currency = NULL, tier = NULL, userComment = NULL)
}
\arguments{
\item{amountDisplayString}{A rendered string that displays the fund amount and currency to the user}

\item{amountMicros}{The amount purchased by the user, in micros (1,750,000 micros = 1}

\item{currency}{The currency in which the purchase was made}

\item{tier}{The tier in which the amount belongs to}

\item{userComment}{The comment added by the user to this Super Chat event}
}
\value{
LiveChatSuperChatDetails object
}
\description{
LiveChatSuperChatDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
