% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatTextMessageDetails}
\alias{LiveChatTextMessageDetails}
\title{LiveChatTextMessageDetails Object}
\usage{
LiveChatTextMessageDetails(messageText = NULL)
}
\arguments{
\item{messageText}{The user's message}
}
\value{
LiveChatTextMessageDetails object
}
\description{
LiveChatTextMessageDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

