% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveChatUserBannedMessageDetails}
\alias{LiveChatUserBannedMessageDetails}
\title{LiveChatUserBannedMessageDetails Object}
\usage{
LiveChatUserBannedMessageDetails(banDurationSeconds = NULL, banType = NULL,
  bannedUserDetails = NULL)
}
\arguments{
\item{banDurationSeconds}{The duration of the ban}

\item{banType}{The type of ban}

\item{bannedUserDetails}{The details of the user that was banned}
}
\value{
LiveChatUserBannedMessageDetails object
}
\description{
LiveChatUserBannedMessageDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

