% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveStream}
\alias{LiveStream}
\title{LiveStream Object}
\usage{
LiveStream(cdn = NULL, contentDetails = NULL, etag = NULL, id = NULL,
  snippet = NULL, status = NULL)
}
\arguments{
\item{cdn}{The cdn object defines the live stream's content delivery network (CDN) settings}

\item{contentDetails}{The content_details object contains information about the stream, including the closed captions ingestion URL}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the stream}

\item{snippet}{The snippet object contains basic details about the stream, including its channel, title, and description}

\item{status}{The status object contains information about live stream's status}
}
\value{
LiveStream object
}
\description{
LiveStream Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A live stream describes a live ingestion point.
}
\seealso{
Other LiveStream functions: \code{\link{liveStreams.insert}},
  \code{\link{liveStreams.update}}
}

