% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveStreamConfigurationIssue}
\alias{LiveStreamConfigurationIssue}
\title{LiveStreamConfigurationIssue Object}
\usage{
LiveStreamConfigurationIssue(description = NULL, reason = NULL,
  severity = NULL, type = NULL)
}
\arguments{
\item{description}{The long-form description of the issue and how to resolve it}

\item{reason}{The short-form reason for this issue}

\item{severity}{How severe this issue is to the stream}

\item{type}{The kind of error happening}
}
\value{
LiveStreamConfigurationIssue object
}
\description{
LiveStreamConfigurationIssue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
