% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveStreamContentDetails}
\alias{LiveStreamContentDetails}
\title{LiveStreamContentDetails Object}
\usage{
LiveStreamContentDetails(closedCaptionsIngestionUrl = NULL,
  isReusable = NULL)
}
\arguments{
\item{closedCaptionsIngestionUrl}{The ingestion URL where the closed captions of this stream are sent}

\item{isReusable}{Indicates whether the stream is reusable, which means that it can be bound to multiple broadcasts}
}
\value{
LiveStreamContentDetails object
}
\description{
LiveStreamContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Detailed settings of a stream.
}
