% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveStreamHealthStatus}
\alias{LiveStreamHealthStatus}
\title{LiveStreamHealthStatus Object}
\usage{
LiveStreamHealthStatus(configurationIssues = NULL,
  lastUpdateTimeSeconds = NULL, status = NULL)
}
\arguments{
\item{configurationIssues}{The configurations issues on this stream}

\item{lastUpdateTimeSeconds}{The last time this status was updated (in seconds)}

\item{status}{The status code of this stream}
}
\value{
LiveStreamHealthStatus object
}
\description{
LiveStreamHealthStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
