% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveStreamSnippet}
\alias{LiveStreamSnippet}
\title{LiveStreamSnippet Object}
\usage{
LiveStreamSnippet(channelId = NULL, description = NULL,
  isDefaultStream = NULL, publishedAt = NULL, title = NULL)
}
\arguments{
\item{channelId}{The ID that YouTube uses to uniquely identify the channel that is transmitting the stream}

\item{description}{The stream's description}

\item{isDefaultStream}{No description}

\item{publishedAt}{The date and time that the stream was created}

\item{title}{The stream's title}
}
\value{
LiveStreamSnippet object
}
\description{
LiveStreamSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

