% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LiveStreamStatus}
\alias{LiveStreamStatus}
\title{LiveStreamStatus Object}
\usage{
LiveStreamStatus(healthStatus = NULL, streamStatus = NULL)
}
\arguments{
\item{healthStatus}{The health status of the stream}

\item{streamStatus}{No description}
}
\value{
LiveStreamStatus object
}
\description{
LiveStreamStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Brief description of the live stream status.
}

