% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LocalizedProperty}
\alias{LocalizedProperty}
\title{LocalizedProperty Object}
\usage{
LocalizedProperty(default = NULL, defaultLanguage = NULL,
  localized = NULL)
}
\arguments{
\item{default}{No description}

\item{defaultLanguage}{The language of the default property}

\item{localized}{No description}
}
\value{
LocalizedProperty object
}
\description{
LocalizedProperty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

