% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{LocalizedString}
\alias{LocalizedString}
\title{LocalizedString Object}
\usage{
LocalizedString(language = NULL, value = NULL)
}
\arguments{
\item{language}{No description}

\item{value}{No description}
}
\value{
LocalizedString object
}
\description{
LocalizedString Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
