% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{MonitorStreamInfo}
\alias{MonitorStreamInfo}
\title{MonitorStreamInfo Object}
\usage{
MonitorStreamInfo(broadcastStreamDelayMs = NULL, embedHtml = NULL,
  enableMonitorStream = NULL)
}
\arguments{
\item{broadcastStreamDelayMs}{If you have set the enableMonitorStream property to true, then this property determines the length of the live broadcast delay}

\item{embedHtml}{HTML code that embeds a player that plays the monitor stream}

\item{enableMonitorStream}{This value determines whether the monitor stream is enabled for the broadcast}
}
\value{
MonitorStreamInfo object
}
\description{
MonitorStreamInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings and Info of the monitor stream
}

