% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PageInfo}
\alias{PageInfo}
\title{PageInfo Object}
\usage{
PageInfo(resultsPerPage = NULL, totalResults = NULL)
}
\arguments{
\item{resultsPerPage}{The number of results included in the API response}

\item{totalResults}{The total number of results in the result set}
}
\value{
PageInfo object
}
\description{
PageInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Paging details for lists of resources, including total number of items available and number of resources returned in a single page.
}
