% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Playlist.localizations}
\alias{Playlist.localizations}
\title{Playlist.localizations Object}
\usage{
Playlist.localizations()
}
\value{
Playlist.localizations object
}
\description{
Playlist.localizations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Localizations for different languages
}
\seealso{
Other Playlist functions: \code{\link{Playlist}},
  \code{\link{playlists.insert}},
  \code{\link{playlists.update}}
}
