% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Playlist}
\alias{Playlist}
\title{Playlist Object}
\usage{
Playlist(Playlist.localizations = NULL, contentDetails = NULL,
  etag = NULL, id = NULL, localizations = NULL, player = NULL,
  snippet = NULL, status = NULL)
}
\arguments{
\item{Playlist.localizations}{The \link{Playlist.localizations} object or list of objects}

\item{contentDetails}{The contentDetails object contains information like video count}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the playlist}

\item{localizations}{Localizations for different languages}

\item{player}{The player object contains information that you would use to play the playlist in an embedded player}

\item{snippet}{The snippet object contains basic details about the playlist, such as its title and description}

\item{status}{The status object contains status information for the playlist}
}
\value{
Playlist object
}
\description{
Playlist Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A playlist resource represents a YouTube playlist. A playlist is a collection of videos that can be viewed sequentially and shared with other users. A playlist can contain up to 200 videos, and YouTube does not limit the number of playlists that each user creates. By default, playlists are publicly visible to other users, but playlists can be public or private.YouTube also uses playlists to identify special collections of videos for a channel, such as:  - uploaded videos - favorite videos - positively rated (liked) videos - watch history - watch later  To be more specific, these lists are associated with a channel, which is a collection of a person, group, or company's videos, playlists, and other YouTube information. You can retrieve the playlist IDs for each of these lists from the  channel resource for a given channel.You can then use the   playlistItems.list method to retrieve any of those lists. You can also add or remove items from those lists by calling the   playlistItems.insert and   playlistItems.delete methods.
}
\seealso{
Other Playlist functions: \code{\link{Playlist.localizations}},
  \code{\link{playlists.insert}},
  \code{\link{playlists.update}}
}
