% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistContentDetails}
\alias{PlaylistContentDetails}
\title{PlaylistContentDetails Object}
\usage{
PlaylistContentDetails(itemCount = NULL)
}
\arguments{
\item{itemCount}{The number of videos in the playlist}
}
\value{
PlaylistContentDetails object
}
\description{
PlaylistContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

