% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistItem}
\alias{PlaylistItem}
\title{PlaylistItem Object}
\usage{
PlaylistItem(contentDetails = NULL, etag = NULL, id = NULL,
  snippet = NULL, status = NULL)
}
\arguments{
\item{contentDetails}{The contentDetails object is included in the resource if the included item is a YouTube video}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the playlist item}

\item{snippet}{The snippet object contains basic details about the playlist item, such as its title and position in the playlist}

\item{status}{The status object contains information about the playlist item's privacy status}
}
\value{
PlaylistItem object
}
\description{
PlaylistItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A playlistItem resource identifies another resource, such as a video, that is included in a playlist. In addition, the playlistItem  resource contains details about the included resource that pertain specifically to how that resource is used in that playlist.YouTube uses playlists to identify special collections of videos for a channel, such as:  - uploaded videos - favorite videos - positively rated (liked) videos - watch history - watch later  To be more specific, these lists are associated with a channel, which is a collection of a person, group, or company's videos, playlists, and other YouTube information.You can retrieve the playlist IDs for each of these lists from the  channel resource  for a given channel. You can then use the   playlistItems.list method to retrieve any of those lists. You can also add or remove items from those lists by calling the   playlistItems.insert and   playlistItems.delete methods. For example, if a user gives a positive rating to a video, you would insert that video into the liked videos playlist for that user's channel.
}
\seealso{
Other PlaylistItem functions: \code{\link{playlistItems.insert}},
  \code{\link{playlistItems.update}}
}
