% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistItemContentDetails}
\alias{PlaylistItemContentDetails}
\title{PlaylistItemContentDetails Object}
\usage{
PlaylistItemContentDetails(endAt = NULL, note = NULL, startAt = NULL,
  videoId = NULL, videoPublishedAt = NULL)
}
\arguments{
\item{endAt}{The time, measured in seconds from the start of the video, when the video should stop playing}

\item{note}{A user-generated note for this item}

\item{startAt}{The time, measured in seconds from the start of the video, when the video should start playing}

\item{videoId}{The ID that YouTube uses to uniquely identify a video}

\item{videoPublishedAt}{The date and time that the video was published to YouTube}
}
\value{
PlaylistItemContentDetails object
}
\description{
PlaylistItemContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
