% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{playlistItems.list}
\alias{playlistItems.list}
\title{Returns a collection of playlist items that match the API request parameters. You can retrieve all of the playlist items in a specified playlist or retrieve one or more playlist items by their unique IDs.}
\usage{
playlistItems.list(part, id = NULL, maxResults = NULL,
  onBehalfOfContentOwner = NULL, pageToken = NULL, playlistId = NULL,
  videoId = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more playlistItem resource properties that the API response will include}

\item{id}{The id parameter specifies a comma-separated list of one or more unique playlist item IDs}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}

\item{playlistId}{The playlistId parameter specifies the unique ID of the playlist for which you want to retrieve playlist items}

\item{videoId}{The videoId parameter specifies that the request should return only the playlist items that contain the specified video}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
