% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{playlistItems.update}
\alias{playlistItems.update}
\title{Modifies a playlist item. For example, you could update the item's position in the playlist.}
\usage{
playlistItems.update(PlaylistItem, part)
}
\arguments{
\item{PlaylistItem}{The \link{PlaylistItem} object to pass to this method}

\item{part}{The part parameter serves two purposes in this operation}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other PlaylistItem functions: \code{\link{PlaylistItem}},
  \code{\link{playlistItems.insert}}
}
