% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistItemStatus}
\alias{PlaylistItemStatus}
\title{PlaylistItemStatus Object}
\usage{
PlaylistItemStatus(privacyStatus = NULL)
}
\arguments{
\item{privacyStatus}{This resource's privacy status}
}
\value{
PlaylistItemStatus object
}
\description{
PlaylistItemStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the playlist item's privacy status.
}

