% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistLocalization}
\alias{PlaylistLocalization}
\title{PlaylistLocalization Object}
\usage{
PlaylistLocalization(description = NULL, title = NULL)
}
\arguments{
\item{description}{The localized strings for playlist's description}

\item{title}{The localized strings for playlist's title}
}
\value{
PlaylistLocalization object
}
\description{
PlaylistLocalization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Playlist localization setting
}

