% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistPlayer}
\alias{PlaylistPlayer}
\title{PlaylistPlayer Object}
\usage{
PlaylistPlayer(embedHtml = NULL)
}
\arguments{
\item{embedHtml}{An <iframe> tag that embeds a player that will play the playlist}
}
\value{
PlaylistPlayer object
}
\description{
PlaylistPlayer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

