% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{playlists.insert}
\alias{playlists.insert}
\title{Creates a playlist.}
\usage{
playlists.insert(Playlist, part, onBehalfOfContentOwner = NULL,
  onBehalfOfContentOwnerChannel = NULL)
}
\arguments{
\item{Playlist}{The \link{Playlist} object to pass to this method}

\item{part}{The part parameter serves two purposes in this operation}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other Playlist functions: \code{\link{Playlist.localizations}},
  \code{\link{Playlist}}, \code{\link{playlists.update}}
}
