% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{playlists.list}
\alias{playlists.list}
\title{Returns a collection of playlists that match the API request parameters. For example, you can retrieve all playlists that the authenticated user owns, or you can retrieve one or more playlists by their unique IDs.}
\usage{
playlists.list(part, channelId = NULL, hl = NULL, id = NULL,
  maxResults = NULL, mine = NULL, onBehalfOfContentOwner = NULL,
  onBehalfOfContentOwnerChannel = NULL, pageToken = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more playlist resource properties that the API response will include}

\item{channelId}{This value indicates that the API should only return the specified channel's playlists}

\item{hl}{The hl parameter should be used for filter out the properties that are not in the given language}

\item{id}{The id parameter specifies a comma-separated list of the YouTube playlist ID(s) for the resource(s) that are being retrieved}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{mine}{Set this parameter's value to true to instruct the API to only return playlists owned by the authenticated user}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

