% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistSnippet}
\alias{PlaylistSnippet}
\title{PlaylistSnippet Object}
\usage{
PlaylistSnippet(channelId = NULL, channelTitle = NULL,
  defaultLanguage = NULL, description = NULL, localized = NULL,
  publishedAt = NULL, tags = NULL, thumbnails = NULL, title = NULL)
}
\arguments{
\item{channelId}{The ID that YouTube uses to uniquely identify the channel that published the playlist}

\item{channelTitle}{The channel title of the channel that the video belongs to}

\item{defaultLanguage}{The language of the playlist's default title and description}

\item{description}{The playlist's description}

\item{localized}{Localized title and description, read-only}

\item{publishedAt}{The date and time that the playlist was created}

\item{tags}{Keyword tags associated with the playlist}

\item{thumbnails}{A map of thumbnail images associated with the playlist}

\item{title}{The playlist's title}
}
\value{
PlaylistSnippet object
}
\description{
PlaylistSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a playlist, including title, description and thumbnails.
}
