% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PlaylistStatus}
\alias{PlaylistStatus}
\title{PlaylistStatus Object}
\usage{
PlaylistStatus(privacyStatus = NULL)
}
\arguments{
\item{privacyStatus}{The playlist's privacy status}
}
\value{
PlaylistStatus object
}
\description{
PlaylistStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
