% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PromotedItemId}
\alias{PromotedItemId}
\title{PromotedItemId Object}
\usage{
PromotedItemId(recentlyUploadedBy = NULL, type = NULL, videoId = NULL,
  websiteUrl = NULL)
}
\arguments{
\item{recentlyUploadedBy}{If type is recentUpload, this field identifies the channel from which to take the recent upload}

\item{type}{Describes the type of the promoted item}

\item{videoId}{If the promoted item represents a video, this field represents the unique YouTube ID identifying it}

\item{websiteUrl}{If the promoted item represents a website, this field represents the url pointing to the website}
}
\value{
PromotedItemId object
}
\description{
PromotedItemId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a single promoted item id. It is a union of various possible types.
}
