% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{PropertyValue}
\alias{PropertyValue}
\title{PropertyValue Object}
\usage{
PropertyValue(property = NULL, value = NULL)
}
\arguments{
\item{property}{A property}

\item{value}{The property's value}
}
\value{
PropertyValue object
}
\description{
PropertyValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A pair Property / Value.
}

