% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ResourceId}
\alias{ResourceId}
\title{ResourceId Object}
\usage{
ResourceId(channelId = NULL, playlistId = NULL, videoId = NULL)
}
\arguments{
\item{channelId}{The ID that YouTube uses to uniquely identify the referred resource, if that resource is a channel}

\item{playlistId}{The ID that YouTube uses to uniquely identify the referred resource, if that resource is a playlist}

\item{videoId}{The ID that YouTube uses to uniquely identify the referred resource, if that resource is a video}
}
\value{
ResourceId object
}
\description{
ResourceId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A resource id is a generic reference that points to another YouTube resource.
}
