% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SearchResult}
\alias{SearchResult}
\title{SearchResult Object}
\usage{
SearchResult(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The id object contains information that can be used to uniquely identify the resource that matches the search request}

\item{snippet}{The snippet object contains basic details about a search result, such as its title or description}
}
\value{
SearchResult object
}
\description{
SearchResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A search result contains information about a YouTube video, channel, or playlist that matches the search parameters specified in an API request. While a search result points to a uniquely identifiable resource, like a video, it does not have its own persistent data.
}

