% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SearchResultSnippet}
\alias{SearchResultSnippet}
\title{SearchResultSnippet Object}
\usage{
SearchResultSnippet(channelId = NULL, channelTitle = NULL,
  description = NULL, liveBroadcastContent = NULL, publishedAt = NULL,
  thumbnails = NULL, title = NULL)
}
\arguments{
\item{channelId}{The value that YouTube uses to uniquely identify the channel that published the resource that the search result identifies}

\item{channelTitle}{The title of the channel that published the resource that the search result identifies}

\item{description}{A description of the search result}

\item{liveBroadcastContent}{It indicates if the resource (video or channel) has upcoming/active live broadcast content}

\item{publishedAt}{The creation date and time of the resource that the search result identifies}

\item{thumbnails}{A map of thumbnail images associated with the search result}

\item{title}{The title of the search result}
}
\value{
SearchResultSnippet object
}
\description{
SearchResultSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a search result, including title, description and thumbnails of the item referenced by the search result.
}
