% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Sponsor}
\alias{Sponsor}
\title{Sponsor Object}
\usage{
Sponsor(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the sponsor}

\item{snippet}{The snippet object contains basic details about the sponsor}
}
\value{
Sponsor object
}
\description{
Sponsor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A sponsor resource represents a sponsor for a YouTube channel. A sponsor provides recurring monetary support to a creator and receives special benefits.
}

