% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{sponsors.list}
\alias{sponsors.list}
\title{Lists sponsors for a channel.}
\usage{
sponsors.list(part, filter = NULL, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{part}{The part parameter specifies the sponsor resource parts that the API response will include}

\item{filter}{The filter parameter specifies which channel sponsors to return}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

