% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SponsorSnippet}
\alias{SponsorSnippet}
\title{SponsorSnippet Object}
\usage{
SponsorSnippet(channelId = NULL, sponsorDetails = NULL,
  sponsorSince = NULL)
}
\arguments{
\item{channelId}{The id of the channel being sponsored}

\item{sponsorDetails}{Details about the sponsor}

\item{sponsorSince}{The date and time when the user became a sponsor}
}
\value{
SponsorSnippet object
}
\description{
SponsorSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
