% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Subscription}
\alias{Subscription}
\title{Subscription Object}
\usage{
Subscription(contentDetails = NULL, etag = NULL, id = NULL,
  snippet = NULL, subscriberSnippet = NULL)
}
\arguments{
\item{contentDetails}{The contentDetails object contains basic statistics about the subscription}

\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the subscription}

\item{snippet}{The snippet object contains basic details about the subscription, including its title and the channel that the user subscribed to}

\item{subscriberSnippet}{The subscriberSnippet object contains basic details about the sbuscriber}
}
\value{
Subscription object
}
\description{
Subscription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A subscription resource contains information about a YouTube user subscription. A subscription notifies a user when new videos are added to a channel or when another user takes one of several actions on YouTube, such as uploading a video, rating a video, or commenting on a video.
}
\seealso{
Other Subscription functions: \code{\link{subscriptions.insert}}
}
