% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SubscriptionContentDetails}
\alias{SubscriptionContentDetails}
\title{SubscriptionContentDetails Object}
\usage{
SubscriptionContentDetails(activityType = NULL, newItemCount = NULL,
  totalItemCount = NULL)
}
\arguments{
\item{activityType}{The type of activity this subscription is for (only uploads, everything)}

\item{newItemCount}{The number of new items in the subscription since its content was last read}

\item{totalItemCount}{The approximate number of items that the subscription points to}
}
\value{
SubscriptionContentDetails object
}
\description{
SubscriptionContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about the content to witch a subscription refers.
}

