% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SubscriptionListResponse}
\alias{SubscriptionListResponse}
\title{SubscriptionListResponse Object}
\usage{
SubscriptionListResponse(etag = NULL, eventId = NULL, items = NULL,
  nextPageToken = NULL, pageInfo = NULL, prevPageToken = NULL,
  tokenPagination = NULL, visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of subscriptions that match the request criteria}

\item{nextPageToken}{The token that can be used as the value of the pageToken parameter to retrieve the next page in the result set}

\item{pageInfo}{No description}

\item{prevPageToken}{The token that can be used as the value of the pageToken parameter to retrieve the previous page in the result set}

\item{tokenPagination}{No description}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
SubscriptionListResponse object
}
\description{
SubscriptionListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

