% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{subscriptions.list}
\alias{subscriptions.list}
\title{Returns subscription resources that match the API request criteria.}
\usage{
subscriptions.list(part, channelId = NULL, forChannelId = NULL, id = NULL,
  maxResults = NULL, mine = NULL, myRecentSubscribers = NULL,
  mySubscribers = NULL, onBehalfOfContentOwner = NULL,
  onBehalfOfContentOwnerChannel = NULL, order = NULL, pageToken = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more subscription resource properties that the API response will include}

\item{channelId}{The channelId parameter specifies a YouTube channel ID}

\item{forChannelId}{The forChannelId parameter specifies a comma-separated list of channel IDs}

\item{id}{The id parameter specifies a comma-separated list of the YouTube subscription ID(s) for the resource(s) that are being retrieved}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{mine}{Set this parameter's value to true to retrieve a feed of the authenticated user's subscriptions}

\item{myRecentSubscribers}{Set this parameter's value to true to retrieve a feed of the subscribers of the authenticated user in reverse chronological order (newest first)}

\item{mySubscribers}{Set this parameter's value to true to retrieve a feed of the subscribers of the authenticated user in no particular order}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}

\item{order}{The order parameter specifies the method that will be used to sort resources in the API response}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
