% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SubscriptionSnippet}
\alias{SubscriptionSnippet}
\title{SubscriptionSnippet Object}
\usage{
SubscriptionSnippet(channelId = NULL, channelTitle = NULL,
  description = NULL, publishedAt = NULL, resourceId = NULL,
  thumbnails = NULL, title = NULL)
}
\arguments{
\item{channelId}{The ID that YouTube uses to uniquely identify the subscriber's channel}

\item{channelTitle}{Channel title for the channel that the subscription belongs to}

\item{description}{The subscription's details}

\item{publishedAt}{The date and time that the subscription was created}

\item{resourceId}{The id object contains information about the channel that the user subscribed to}

\item{thumbnails}{A map of thumbnail images associated with the video}

\item{title}{The subscription's title}
}
\value{
SubscriptionSnippet object
}
\description{
SubscriptionSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a subscription, including title, description and thumbnails of the subscribed item.
}

