% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SubscriptionSubscriberSnippet}
\alias{SubscriptionSubscriberSnippet}
\title{SubscriptionSubscriberSnippet Object}
\usage{
SubscriptionSubscriberSnippet(channelId = NULL, description = NULL,
  thumbnails = NULL, title = NULL)
}
\arguments{
\item{channelId}{The channel ID of the subscriber}

\item{description}{The description of the subscriber}

\item{thumbnails}{Thumbnails for this subscriber}

\item{title}{The title of the subscriber}
}
\value{
SubscriptionSubscriberSnippet object
}
\description{
SubscriptionSubscriberSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a subscription's subscriber including title, description, channel ID and thumbnails.
}

