% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SuperChatEvent}
\alias{SuperChatEvent}
\title{SuperChatEvent Object}
\usage{
SuperChatEvent(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube assigns to uniquely identify the Super Chat event}

\item{snippet}{The snippet object contains basic details about the Super Chat event}
}
\value{
SuperChatEvent object
}
\description{
SuperChatEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A superChatEvent resource represents a Super Chat purchase on a YouTube channel.
}
