% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{SuperChatEventSnippet}
\alias{SuperChatEventSnippet}
\title{SuperChatEventSnippet Object}
\usage{
SuperChatEventSnippet(amountMicros = NULL, channelId = NULL,
  commentText = NULL, createdAt = NULL, currency = NULL,
  displayString = NULL, messageType = NULL, supporterDetails = NULL)
}
\arguments{
\item{amountMicros}{The purchase amount, in micros of the purchase currency}

\item{channelId}{Channel id where the event occurred}

\item{commentText}{The text contents of the comment left by the user}

\item{createdAt}{The date and time when the event occurred}

\item{currency}{The currency in which the purchase was made}

\item{displayString}{A rendered string that displays the purchase amount and currency (e}

\item{messageType}{The tier for the paid message, which is based on the amount of money spent to purchase the message}

\item{supporterDetails}{Details about the supporter}
}
\value{
SuperChatEventSnippet object
}
\description{
SuperChatEventSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
