% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Thumbnail}
\alias{Thumbnail}
\title{Thumbnail Object}
\usage{
Thumbnail(height = NULL, url = NULL, width = NULL)
}
\arguments{
\item{height}{(Optional) Height of the thumbnail image}

\item{url}{The thumbnail image's URL}

\item{width}{(Optional) Width of the thumbnail image}
}
\value{
Thumbnail object
}
\description{
Thumbnail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A thumbnail is an image representing a YouTube resource.
}
