% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{ThumbnailDetails}
\alias{ThumbnailDetails}
\title{ThumbnailDetails Object}
\usage{
ThumbnailDetails(default = NULL, high = NULL, maxres = NULL,
  medium = NULL, standard = NULL)
}
\arguments{
\item{default}{The default image for this resource}

\item{high}{The high quality image for this resource}

\item{maxres}{The maximum resolution quality image for this resource}

\item{medium}{The medium quality image for this resource}

\item{standard}{The standard quality image for this resource}
}
\value{
ThumbnailDetails object
}
\description{
ThumbnailDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Internal representation of thumbnails for a YouTube resource.
}
