% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{thumbnails.set}
\alias{thumbnails.set}
\title{Uploads a custom video thumbnail to YouTube and sets it for a video.}
\usage{
thumbnails.set(videoId, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{videoId}{The videoId parameter specifies a YouTube video ID for which the custom video thumbnail is being provided}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.upload
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.upload, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}

