% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{TokenPagination}
\alias{TokenPagination}
\title{TokenPagination Object}
\usage{
TokenPagination()
}
\value{
TokenPagination object
}
\description{
TokenPagination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stub token pagination template to suppress results.
}

