% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Video.localizations}
\alias{Video.localizations}
\title{Video.localizations Object}
\usage{
Video.localizations()
}
\value{
Video.localizations object
}
\description{
Video.localizations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List with all localizations.
}
\seealso{
Other Video functions: \code{\link{Video}},
  \code{\link{videos.insert}}, \code{\link{videos.update}}
}
