% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{Video}
\alias{Video}
\title{Video Object}
\usage{
Video(Video.localizations = NULL, ageGating = NULL, contentDetails = NULL,
  etag = NULL, fileDetails = NULL, id = NULL,
  liveStreamingDetails = NULL, localizations = NULL,
  monetizationDetails = NULL, player = NULL, processingDetails = NULL,
  projectDetails = NULL, recordingDetails = NULL, snippet = NULL,
  statistics = NULL, status = NULL, suggestions = NULL,
  topicDetails = NULL)
}
\arguments{
\item{Video.localizations}{The \link{Video.localizations} object or list of objects}

\item{ageGating}{Age restriction details related to a video}

\item{contentDetails}{The contentDetails object contains information about the video content, including the length of the video and its aspect ratio}

\item{etag}{Etag of this resource}

\item{fileDetails}{The fileDetails object encapsulates information about the video file that was uploaded to YouTube, including the file's resolution, duration, audio and video codecs, stream bitrates, and more}

\item{id}{The ID that YouTube uses to uniquely identify the video}

\item{liveStreamingDetails}{The liveStreamingDetails object contains metadata about a live video broadcast}

\item{localizations}{List with all localizations}

\item{monetizationDetails}{The monetizationDetails object encapsulates information about the monetization status of the video}

\item{player}{The player object contains information that you would use to play the video in an embedded player}

\item{processingDetails}{The processingProgress object encapsulates information about YouTube's progress in processing the uploaded video file}

\item{projectDetails}{The projectDetails object contains information about the project specific video metadata}

\item{recordingDetails}{The recordingDetails object encapsulates information about the location, date and address where the video was recorded}

\item{snippet}{The snippet object contains basic details about the video, such as its title, description, and category}

\item{statistics}{The statistics object contains statistics about the video}

\item{status}{The status object contains information about the video's uploading, processing, and privacy statuses}

\item{suggestions}{The suggestions object encapsulates suggestions that identify opportunities to improve the video quality or the metadata for the uploaded video}

\item{topicDetails}{The topicDetails object encapsulates information about Freebase topics associated with the video}
}
\value{
Video object
}
\description{
Video Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A video resource represents a YouTube video.
}
\seealso{
Other Video functions: \code{\link{Video.localizations}},
  \code{\link{videos.insert}}, \code{\link{videos.update}}
}

