% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoAbuseReport}
\alias{VideoAbuseReport}
\title{VideoAbuseReport Object}
\usage{
VideoAbuseReport(comments = NULL, language = NULL, reasonId = NULL,
  secondaryReasonId = NULL, videoId = NULL)
}
\arguments{
\item{comments}{Additional comments regarding the abuse report}

\item{language}{The language that the content was viewed in}

\item{reasonId}{The high-level, or primary, reason that the content is abusive}

\item{secondaryReasonId}{The specific, or secondary, reason that this content is abusive (if available)}

\item{videoId}{The ID that YouTube uses to uniquely identify the video}
}
\value{
VideoAbuseReport object
}
\description{
VideoAbuseReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other VideoAbuseReport functions: \code{\link{videos.reportAbuse}}
}
