% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoAbuseReportReason}
\alias{VideoAbuseReportReason}
\title{VideoAbuseReportReason Object}
\usage{
VideoAbuseReportReason(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID of this abuse report reason}

\item{snippet}{The snippet object contains basic details about the abuse report reason}
}
\value{
VideoAbuseReportReason object
}
\description{
VideoAbuseReportReason Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A videoAbuseReportReason resource identifies a reason that a video could be reported as abusive. Video abuse report reasons are used with video.ReportAbuse.
}
