% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoAbuseReportReasonListResponse}
\alias{VideoAbuseReportReasonListResponse}
\title{VideoAbuseReportReasonListResponse Object}
\usage{
VideoAbuseReportReasonListResponse(etag = NULL, eventId = NULL,
  items = NULL, visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of valid abuse reasons that are used with video}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
VideoAbuseReportReasonListResponse object
}
\description{
VideoAbuseReportReasonListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
