% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoAbuseReportReasonSnippet}
\alias{VideoAbuseReportReasonSnippet}
\title{VideoAbuseReportReasonSnippet Object}
\usage{
VideoAbuseReportReasonSnippet(label = NULL, secondaryReasons = NULL)
}
\arguments{
\item{label}{The localized label belonging to this abuse report reason}

\item{secondaryReasons}{The secondary reasons associated with this reason, if any are available}
}
\value{
VideoAbuseReportReasonSnippet object
}
\description{
VideoAbuseReportReasonSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a video category, such as its localized title.
}
