% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoAbuseReportSecondaryReason}
\alias{VideoAbuseReportSecondaryReason}
\title{VideoAbuseReportSecondaryReason Object}
\usage{
VideoAbuseReportSecondaryReason(id = NULL, label = NULL)
}
\arguments{
\item{id}{The ID of this abuse report secondary reason}

\item{label}{The localized label for this abuse report secondary reason}
}
\value{
VideoAbuseReportSecondaryReason object
}
\description{
VideoAbuseReportSecondaryReason Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
