% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoAgeGating}
\alias{VideoAgeGating}
\title{VideoAgeGating Object}
\usage{
VideoAgeGating(alcoholContent = NULL, restricted = NULL,
  videoGameRating = NULL)
}
\arguments{
\item{alcoholContent}{Indicates whether or not the video has alcoholic beverage content}

\item{restricted}{Age-restricted trailers}

\item{videoGameRating}{Video game rating, if any}
}
\value{
VideoAgeGating object
}
\description{
VideoAgeGating Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

