% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoCategory}
\alias{VideoCategory}
\title{VideoCategory Object}
\usage{
VideoCategory(etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{id}{The ID that YouTube uses to uniquely identify the video category}

\item{snippet}{The snippet object contains basic details about the video category, including its title}
}
\value{
VideoCategory object
}
\description{
VideoCategory Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A videoCategory resource identifies a category that has been or could be associated with uploaded videos.
}

