% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoCategorySnippet}
\alias{VideoCategorySnippet}
\title{VideoCategorySnippet Object}
\usage{
VideoCategorySnippet(assignable = NULL, channelId = NULL, title = NULL)
}
\arguments{
\item{assignable}{No description}

\item{channelId}{The YouTube channel that created the video category}

\item{title}{The video category's title}
}
\value{
VideoCategorySnippet object
}
\description{
VideoCategorySnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a video category, such as its localized title.
}
