% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoContentDetails}
\alias{VideoContentDetails}
\title{VideoContentDetails Object}
\usage{
VideoContentDetails(caption = NULL, contentRating = NULL,
  countryRestriction = NULL, definition = NULL, dimension = NULL,
  duration = NULL, hasCustomThumbnail = NULL, licensedContent = NULL,
  projection = NULL, regionRestriction = NULL)
}
\arguments{
\item{caption}{The value of captions indicates whether the video has captions or not}

\item{contentRating}{Specifies the ratings that the video received under various rating schemes}

\item{countryRestriction}{The countryRestriction object contains information about the countries where a video is (or is not) viewable}

\item{definition}{The value of definition indicates whether the video is available in high definition or only in standard definition}

\item{dimension}{The value of dimension indicates whether the video is available in 3D or in 2D}

\item{duration}{The length of the video}

\item{hasCustomThumbnail}{Indicates whether the video uploader has provided a custom thumbnail image for the video}

\item{licensedContent}{The value of is_license_content indicates whether the video is licensed content}

\item{projection}{Specifies the projection format of the video}

\item{regionRestriction}{The regionRestriction object contains information about the countries where a video is (or is not) viewable}
}
\value{
VideoContentDetails object
}
\description{
VideoContentDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about the content of a YouTube Video.
}
