% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoContentDetailsRegionRestriction}
\alias{VideoContentDetailsRegionRestriction}
\title{VideoContentDetailsRegionRestriction Object}
\usage{
VideoContentDetailsRegionRestriction(allowed = NULL, blocked = NULL)
}
\arguments{
\item{allowed}{A list of region codes that identify countries where the video is viewable}

\item{blocked}{A list of region codes that identify countries where the video is blocked}
}
\value{
VideoContentDetailsRegionRestriction object
}
\description{
VideoContentDetailsRegionRestriction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
DEPRECATED Region restriction of the video.
}

