% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoFileDetails}
\alias{VideoFileDetails}
\title{VideoFileDetails Object}
\usage{
VideoFileDetails(audioStreams = NULL, bitrateBps = NULL, container = NULL,
  creationTime = NULL, durationMs = NULL, fileName = NULL,
  fileSize = NULL, fileType = NULL, videoStreams = NULL)
}
\arguments{
\item{audioStreams}{A list of audio streams contained in the uploaded video file}

\item{bitrateBps}{The uploaded video file's combined (video and audio) bitrate in bits per second}

\item{container}{The uploaded video file's container format}

\item{creationTime}{The date and time when the uploaded video file was created}

\item{durationMs}{The length of the uploaded video in milliseconds}

\item{fileName}{The uploaded file's name}

\item{fileSize}{The uploaded file's size in bytes}

\item{fileType}{The uploaded file's type as detected by YouTube's video processing engine}

\item{videoStreams}{A list of video streams contained in the uploaded video file}
}
\value{
VideoFileDetails object
}
\description{
VideoFileDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes original video file properties, including technical details about audio and video streams, but also metadata information like content length, digitization time, or geotagging information.
}
