% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoFileDetailsAudioStream}
\alias{VideoFileDetailsAudioStream}
\title{VideoFileDetailsAudioStream Object}
\usage{
VideoFileDetailsAudioStream(bitrateBps = NULL, channelCount = NULL,
  codec = NULL, vendor = NULL)
}
\arguments{
\item{bitrateBps}{The audio stream's bitrate, in bits per second}

\item{channelCount}{The number of audio channels that the stream contains}

\item{codec}{The audio codec that the stream uses}

\item{vendor}{A value that uniquely identifies a video vendor}
}
\value{
VideoFileDetailsAudioStream object
}
\description{
VideoFileDetailsAudioStream Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about an audio stream.
}
