% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoGetRatingResponse}
\alias{VideoGetRatingResponse}
\title{VideoGetRatingResponse Object}
\usage{
VideoGetRatingResponse(etag = NULL, eventId = NULL, items = NULL,
  visitorId = NULL)
}
\arguments{
\item{etag}{Etag of this resource}

\item{eventId}{Serialized EventId of the request which produced this response}

\item{items}{A list of ratings that match the request criteria}

\item{visitorId}{The visitorId identifies the visitor}
}
\value{
VideoGetRatingResponse object
}
\description{
VideoGetRatingResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

