% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoLiveStreamingDetails}
\alias{VideoLiveStreamingDetails}
\title{VideoLiveStreamingDetails Object}
\usage{
VideoLiveStreamingDetails(activeLiveChatId = NULL, actualEndTime = NULL,
  actualStartTime = NULL, concurrentViewers = NULL,
  scheduledEndTime = NULL, scheduledStartTime = NULL)
}
\arguments{
\item{activeLiveChatId}{The ID of the currently active live chat attached to this video}

\item{actualEndTime}{The time that the broadcast actually ended}

\item{actualStartTime}{The time that the broadcast actually started}

\item{concurrentViewers}{The number of viewers currently watching the broadcast}

\item{scheduledEndTime}{The time that the broadcast is scheduled to end}

\item{scheduledStartTime}{The time that the broadcast is scheduled to begin}
}
\value{
VideoLiveStreamingDetails object
}
\description{
VideoLiveStreamingDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about the live streaming metadata.
}
