% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoLocalization}
\alias{VideoLocalization}
\title{VideoLocalization Object}
\usage{
VideoLocalization(description = NULL, title = NULL)
}
\arguments{
\item{description}{Localized version of the video's description}

\item{title}{Localized version of the video's title}
}
\value{
VideoLocalization object
}
\description{
VideoLocalization Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Localized versions of certain video properties (e.g. title).
}

