% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoMonetizationDetails}
\alias{VideoMonetizationDetails}
\title{VideoMonetizationDetails Object}
\usage{
VideoMonetizationDetails(access = NULL)
}
\arguments{
\item{access}{The value of access indicates whether the video can be monetized or not}
}
\value{
VideoMonetizationDetails object
}
\description{
VideoMonetizationDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details about monetization of a YouTube Video.
}

