% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoPlayer}
\alias{VideoPlayer}
\title{VideoPlayer Object}
\usage{
VideoPlayer(embedHeight = NULL, embedHtml = NULL, embedWidth = NULL)
}
\arguments{
\item{embedHeight}{No description}

\item{embedHtml}{An <iframe> tag that embeds a player that will play the video}

\item{embedWidth}{The embed width}
}
\value{
VideoPlayer object
}
\description{
VideoPlayer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Player to be used for a video playback.
}
