% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoProcessingDetails}
\alias{VideoProcessingDetails}
\title{VideoProcessingDetails Object}
\usage{
VideoProcessingDetails(editorSuggestionsAvailability = NULL,
  fileDetailsAvailability = NULL, processingFailureReason = NULL,
  processingIssuesAvailability = NULL, processingProgress = NULL,
  processingStatus = NULL, tagSuggestionsAvailability = NULL,
  thumbnailsAvailability = NULL)
}
\arguments{
\item{editorSuggestionsAvailability}{This value indicates whether video editing suggestions, which might improve video quality or the playback experience, are available for the video}

\item{fileDetailsAvailability}{This value indicates whether file details are available for the uploaded video}

\item{processingFailureReason}{The reason that YouTube failed to process the video}

\item{processingIssuesAvailability}{This value indicates whether the video processing engine has generated suggestions that might improve YouTube's ability to process the the video, warnings that explain video processing problems, or errors that cause video processing problems}

\item{processingProgress}{The processingProgress object contains information about the progress YouTube has made in processing the video}

\item{processingStatus}{The video's processing status}

\item{tagSuggestionsAvailability}{This value indicates whether keyword (tag) suggestions are available for the video}

\item{thumbnailsAvailability}{This value indicates whether thumbnail images have been generated for the video}
}
\value{
VideoProcessingDetails object
}
\description{
VideoProcessingDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes processing status and progress and availability of some other Video resource parts.
}

