% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoProcessingDetailsProcessingProgress}
\alias{VideoProcessingDetailsProcessingProgress}
\title{VideoProcessingDetailsProcessingProgress Object}
\usage{
VideoProcessingDetailsProcessingProgress(partsProcessed = NULL,
  partsTotal = NULL, timeLeftMs = NULL)
}
\arguments{
\item{partsProcessed}{The number of parts of the video that YouTube has already processed}

\item{partsTotal}{An estimate of the total number of parts that need to be processed for the video}

\item{timeLeftMs}{An estimate of the amount of time, in millseconds, that YouTube needs to finish processing the video}
}
\value{
VideoProcessingDetailsProcessingProgress object
}
\description{
VideoProcessingDetailsProcessingProgress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Video processing progress and completion time estimate.
}
