% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoProjectDetails}
\alias{VideoProjectDetails}
\title{VideoProjectDetails Object}
\usage{
VideoProjectDetails(tags = NULL)
}
\arguments{
\item{tags}{A list of project tags associated with the video during the upload}
}
\value{
VideoProjectDetails object
}
\description{
VideoProjectDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Project specific details about the content of a YouTube Video.
}
