% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoRating}
\alias{VideoRating}
\title{VideoRating Object}
\usage{
VideoRating(rating = NULL, videoId = NULL)
}
\arguments{
\item{rating}{No description}

\item{videoId}{No description}
}
\value{
VideoRating object
}
\description{
VideoRating Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
