% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoRecordingDetails}
\alias{VideoRecordingDetails}
\title{VideoRecordingDetails Object}
\usage{
VideoRecordingDetails(location = NULL, locationDescription = NULL,
  recordingDate = NULL)
}
\arguments{
\item{location}{The geolocation information associated with the video}

\item{locationDescription}{The text description of the location where the video was recorded}

\item{recordingDate}{The date and time when the video was recorded}
}
\value{
VideoRecordingDetails object
}
\description{
VideoRecordingDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Recording information associated with the video.
}
