% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{videos.insert}
\alias{videos.insert}
\title{Uploads a video to YouTube and optionally sets the video's metadata.}
\usage{
videos.insert(Video, part, autoLevels = NULL, notifySubscribers = NULL,
  onBehalfOfContentOwner = NULL, onBehalfOfContentOwnerChannel = NULL,
  stabilize = NULL)
}
\arguments{
\item{Video}{The \link{Video} object to pass to this method}

\item{part}{The part parameter serves two purposes in this operation}

\item{autoLevels}{The autoLevels parameter indicates whether YouTube should automatically enhance the video's lighting and color}

\item{notifySubscribers}{The notifySubscribers parameter indicates whether YouTube should send a notification about the new video to users who subscribe to the video's channel}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{onBehalfOfContentOwnerChannel}{This parameter can only be used in a properly authorized request}

\item{stabilize}{The stabilize parameter indicates whether YouTube should adjust the video to remove shaky camera motions}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.upload
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.upload, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other Video functions: \code{\link{Video.localizations}},
  \code{\link{Video}}, \code{\link{videos.update}}
}

