% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{videos.list}
\alias{videos.list}
\title{Returns a list of videos that match the API request parameters.}
\usage{
videos.list(part, chart = NULL, hl = NULL, id = NULL, locale = NULL,
  maxResults = NULL, myRating = NULL, onBehalfOfContentOwner = NULL,
  pageToken = NULL, regionCode = NULL, videoCategoryId = NULL)
}
\arguments{
\item{part}{The part parameter specifies a comma-separated list of one or more video resource properties that the API response will include}

\item{chart}{The chart parameter identifies the chart that you want to retrieve}

\item{hl}{The hl parameter instructs the API to retrieve localized resource metadata for a specific application language that the YouTube website supports}

\item{id}{The id parameter specifies a comma-separated list of the YouTube video ID(s) for the resource(s) that are being retrieved}

\item{locale}{DEPRECATED}

\item{maxResults}{The maxResults parameter specifies the maximum number of items that should be returned in the result set}

\item{myRating}{Set this parameter's value to like or dislike to instruct the API to only return videos liked or disliked by the authenticated user}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}

\item{pageToken}{The pageToken parameter identifies a specific page in the result set that should be returned}

\item{regionCode}{The regionCode parameter instructs the API to select a video chart available in the specified region}

\item{videoCategoryId}{The videoCategoryId parameter identifies the video category for which the chart should be retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}
}
