% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoSnippet}
\alias{VideoSnippet}
\title{VideoSnippet Object}
\usage{
VideoSnippet(categoryId = NULL, channelId = NULL, channelTitle = NULL,
  defaultAudioLanguage = NULL, defaultLanguage = NULL, description = NULL,
  liveBroadcastContent = NULL, localized = NULL, publishedAt = NULL,
  tags = NULL, thumbnails = NULL, title = NULL)
}
\arguments{
\item{categoryId}{The YouTube video category associated with the video}

\item{channelId}{The ID that YouTube uses to uniquely identify the channel that the video was uploaded to}

\item{channelTitle}{Channel title for the channel that the video belongs to}

\item{defaultAudioLanguage}{The default_audio_language property specifies the language spoken in the video's default audio track}

\item{defaultLanguage}{The language of the videos's default snippet}

\item{description}{The video's description}

\item{liveBroadcastContent}{Indicates if the video is an upcoming/active live broadcast}

\item{localized}{Localized snippet selected with the hl parameter}

\item{publishedAt}{The date and time that the video was uploaded}

\item{tags}{A list of keyword tags associated with the video}

\item{thumbnails}{A map of thumbnail images associated with the video}

\item{title}{The video's title}
}
\value{
VideoSnippet object
}
\description{
VideoSnippet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a video, including title, description, uploader, thumbnails and category.
}
