% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoStatistics}
\alias{VideoStatistics}
\title{VideoStatistics Object}
\usage{
VideoStatistics(commentCount = NULL, dislikeCount = NULL,
  favoriteCount = NULL, likeCount = NULL, viewCount = NULL)
}
\arguments{
\item{commentCount}{The number of comments for the video}

\item{dislikeCount}{The number of users who have indicated that they disliked the video by giving it a negative rating}

\item{favoriteCount}{The number of users who currently have the video marked as a favorite video}

\item{likeCount}{The number of users who have indicated that they liked the video by giving it a positive rating}

\item{viewCount}{The number of times the video has been viewed}
}
\value{
VideoStatistics object
}
\description{
VideoStatistics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Statistics about the video, such as the number of times the video was viewed or liked.
}

