% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoStatus}
\alias{VideoStatus}
\title{VideoStatus Object}
\usage{
VideoStatus(embeddable = NULL, failureReason = NULL, license = NULL,
  privacyStatus = NULL, publicStatsViewable = NULL, publishAt = NULL,
  rejectionReason = NULL, uploadStatus = NULL)
}
\arguments{
\item{embeddable}{This value indicates if the video can be embedded on another website}

\item{failureReason}{This value explains why a video failed to upload}

\item{license}{The video's license}

\item{privacyStatus}{The video's privacy status}

\item{publicStatsViewable}{This value indicates if the extended video statistics on the watch page can be viewed by everyone}

\item{publishAt}{The date and time when the video is scheduled to publish}

\item{rejectionReason}{This value explains why YouTube rejected an uploaded video}

\item{uploadStatus}{The status of the uploaded video}
}
\value{
VideoStatus object
}
\description{
VideoStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Basic details about a video category, such as its localized title.
}
