% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoSuggestions}
\alias{VideoSuggestions}
\title{VideoSuggestions Object}
\usage{
VideoSuggestions(editorSuggestions = NULL, processingErrors = NULL,
  processingHints = NULL, processingWarnings = NULL,
  tagSuggestions = NULL)
}
\arguments{
\item{editorSuggestions}{A list of video editing operations that might improve the video quality or playback experience of the uploaded video}

\item{processingErrors}{A list of errors that will prevent YouTube from successfully processing the uploaded video video}

\item{processingHints}{A list of suggestions that may improve YouTube's ability to process the video}

\item{processingWarnings}{A list of reasons why YouTube may have difficulty transcoding the uploaded video or that might result in an erroneous transcoding}

\item{tagSuggestions}{A list of keyword tags that could be added to the video's metadata to increase the likelihood that users will locate your video when searching or browsing on YouTube}
}
\value{
VideoSuggestions object
}
\description{
VideoSuggestions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies suggestions on how to improve video content, including encoding hints, tag suggestions, and editor suggestions.
}
