% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoSuggestionsTagSuggestion}
\alias{VideoSuggestionsTagSuggestion}
\title{VideoSuggestionsTagSuggestion Object}
\usage{
VideoSuggestionsTagSuggestion(categoryRestricts = NULL, tag = NULL)
}
\arguments{
\item{categoryRestricts}{A set of video categories for which the tag is relevant}

\item{tag}{The keyword tag suggested for the video}
}
\value{
VideoSuggestionsTagSuggestion object
}
\description{
VideoSuggestionsTagSuggestion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single tag suggestion with it's relevance information.
}

