% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoTopicDetails}
\alias{VideoTopicDetails}
\title{VideoTopicDetails Object}
\usage{
VideoTopicDetails(relevantTopicIds = NULL, topicIds = NULL)
}
\arguments{
\item{relevantTopicIds}{Similar to topic_id, except that these topics are merely relevant to the video}

\item{topicIds}{A list of Freebase topic IDs that are centrally associated with the video}
}
\value{
VideoTopicDetails object
}
\description{
VideoTopicDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Freebase topic information related to the video.
}
