% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{WatchSettings}
\alias{WatchSettings}
\title{WatchSettings Object}
\usage{
WatchSettings(backgroundColor = NULL, featuredPlaylistId = NULL,
  textColor = NULL)
}
\arguments{
\item{backgroundColor}{The text color for the video watch page's branded area}

\item{featuredPlaylistId}{An ID that uniquely identifies a playlist that displays next to the video player}

\item{textColor}{The background color for the video watch page's branded area}
}
\value{
WatchSettings object
}
\description{
WatchSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Branding properties for the watch. All deprecated.
}
