% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\name{watermarks.set}
\alias{watermarks.set}
\title{Uploads a watermark image to YouTube and sets it for a channel.}
\usage{
watermarks.set(InvideoBranding, channelId, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{InvideoBranding}{The \link{InvideoBranding} object to pass to this method}

\item{channelId}{The channelId parameter specifies the YouTube channel ID for which the watermark is being provided}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.upload
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.force-ssl, https://www.googleapis.com/auth/youtube.upload, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/youtube/v3}{Google Documentation}

Other InvideoBranding functions: \code{\link{InvideoBranding}}
}
