% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_functions.R
\docType{package}
\name{youtube_googleAuthR}
\alias{youtube_googleAuthR}
\alias{youtube_googleAuthR-package}
\title{YouTube Data API
Supports core YouTube features, such as uploading videos, creating and managing playlists, searching for content, and much more.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:22:46
filename: /Users/mark/dev/R/autoGoogleAPI/googleyoutubev3.auto/R/youtube_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.force-ssl
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/youtube.upload
\item https://www.googleapis.com/auth/youtubepartner
\item https://www.googleapis.com/auth/youtubepartner-channel-audit
}
}
