% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytd_data_auth.R
\name{ytd_auth}
\alias{ytd_auth}
\title{Authenticate with YouTube Data API}
\usage{
ytd_auth(clientID, clientSecret)
}
\arguments{
\item{clientID}{The clientID, taken from your credentials in the Google API Console}

\item{clientSecret}{The client secret, taken from your credentials in the Google API Console}
}
\value{
Invisibly returns the oauth file using gar_auth() from the  googleAuthR package.
}
\description{
Authenticate with YouTube Data API
}
\examples{
yt_data_auth(clientID = "my_client_id", clientSecret = "my_client_secret")

}
