% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytd_get_videos.R
\name{ytd_get_videos}
\alias{ytd_get_videos}
\title{Search for videos that match your parameters}
\usage{
ytd_get_videos(part = NULL, chart = NULL, id = NULL, locale = NULL,
  maxHeight = NULL, maxResults = NULL, maxWidth = NULL, myRating = NULL,
  onBehalfOfContentOwner = NULL, pageToken = NULL, regionCode = NULL,
  videoCategoryId = NULL)
}
\arguments{
\item{part}{The part of the video resource you wish to return.}

\item{chart}{The chart you want to retrieve. Currently only "mostPopular" is accepted.}

\item{id}{The YouTube video id for the resource being retrieved.}

\item{locale}{}

\item{maxResults}{The maximum nubmer of results to be returned in the results set.}

\item{maxWidth}{The maximum width of the embedded player returned in the player.embedHtml property}

\item{myRating}{Set this parameter's value to like or dislike to instruct the API to only return videos liked or disliked by the authenticated user.}

\item{onBehalfOfContentOwner}{Indicates that the request's authorization credentials identify a YouTube CMS user who is acting on behalf of the content owner specified in the parameter value}

\item{pageToken}{Identifies a specific page in the result set that should be returned}

\item{regionCode}{Instructs the API to select a video chart available in the specified region}

\item{videoCategoryId}{Identifies the video category for which the chart should be retrieved}

\item{hl}{Instructs the API to retrieve localised resource metadata for a specific application language that the YouTube website supports. The parameter value must be a language code included in the list returned by the i18nLanguages.list method.}
}
\value{
Returns a json response from the YouTube data API.
}
\description{
Search for videos that match your parameters
}
\examples{
yt_get_video(id = "9JwRfNHbBPk", part = "snippet")

Refer to https://developers.google.com/youtube/v3/docs/videos/ for API documentation for resource requests for videos.

}
