% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_objects.R
\name{VideoPlayer}
\alias{VideoPlayer}
\title{VideoPlayer Object}
\usage{
VideoPlayer(embedHtml = NULL)
}
\arguments{
\item{embedHtml}{An <iframe> tag that embeds a player that will play the video}
}
\value{
VideoPlayer object
}
\description{
VideoPlayer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Player to be used for a video playback.
}

